@@ -11,12 +11,12 @@ if TYPE_CHECKING:
     from sphinx.ext.autodoc import Options
 
 
-def autodoc_skip_member(
+def autodoc_skip_member(  # noqa: PLR0917
     app: Sphinx,
     what: Literal["module", "class", "exception", "function", "method", "attribute"],
     name: str,
     obj: object,
-    skip: bool,
+    skip: bool,  # noqa: FBT001
     options: Options,
 ):
     if what == "method" and getattr(obj, "__isabstractmethod__", False):
