@@ -7,29 +7,12 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from typing import Any
 
-try:  # See https://github.com/maresb/hatch-vcs-footgun-example
-    from setuptools_scm import get_version
-
-    __version__ = get_version(root="../..", relative_to=__file__)
-except (ImportError, LookupError):
-    try:
-        from ._version import __version__
-    except ModuleNotFoundError:
-        raise RuntimeError(
-            "anndata is not correctly installed. Please install it, e.g. with pip."
-        )
-
-# Allowing notes to be added to exceptions. See: https://github.com/scverse/anndata/issues/868
-import sys
-
-if sys.version_info < (3, 11):
-    # Backport package for exception groups
-    import exceptiongroup  # noqa: F401
-
 from ._core.anndata import AnnData
+from ._core.extensions import register_anndata_namespace
 from ._core.merge import concat
 from ._core.raw import Raw
 from ._settings import settings
+from ._version import __version__
 from ._warnings import (
     ExperimentalFeatureWarning,
     ImplicitModificationWarning,
@@ -40,22 +23,10 @@ from .io import read_h5ad, read_zarr
 from .utils import module_get_attr_redirect
 
 # Submodules need to be imported last
-from . import abc, experimental, typing, io  # noqa: E402 isort: skip
+from . import abc, experimental, typing, io, types  # isort: skip
 
 # We use these in tests by attribute access
-from . import logging  # noqa: F401, E402 isort: skip
-
-
-def read(*args, **kwargs):
-    import warnings
-
-    warnings.warn(
-        "`anndata.read` is deprecated, use `anndata.read_h5ad` instead. "
-        "`ad.read` will be removed in mid 2024.",
-        FutureWarning,
-    )
-    return read_h5ad(*args, **kwargs)
-
+from . import logging  # noqa: F401  # isort: skip
 
 _DEPRECATED_IO = (
     "read_loom",
@@ -66,7 +37,7 @@ _DEPRECATED_IO = (
     "read_text",
     "read_mtx",
 )
-_DEPRECATED = dict((method, f"io.{method}") for method in _DEPRECATED_IO)
+_DEPRECATED = {method: f"io.{method}" for method in _DEPRECATED_IO}
 
 
 def __getattr__(attr_name: str) -> Any:
@@ -74,25 +45,21 @@ def __getattr__(attr_name: str) -> Any:
 
 
 __all__ = [
-    # Attributes
-    "__version__",
-    "settings",
-    # Submodules
-    "abc",
-    "experimental",
-    "typing",
-    "io",
-    # Classes
     "AnnData",
-    "Raw",
-    # Functions
-    "concat",
-    "read_zarr",
-    "read_h5ad",
-    "read",
-    # Warnings
-    "OldFormatWarning",
-    "WriteWarning",
-    "ImplicitModificationWarning",
     "ExperimentalFeatureWarning",
+    "ImplicitModificationWarning",
+    "OldFormatWarning",
+    "Raw",
+    "WriteWarning",
+    "__version__",
+    "abc",
+    "concat",
+    "experimental",
+    "io",
+    "read_h5ad",
+    "read_zarr",
+    "register_anndata_namespace",
+    "settings",
+    "types",
+    "typing",
 ]
