@@ -8,10 +8,10 @@ import warnings
 from collections import OrderedDict
 from collections.abc import Mapping, MutableMapping, Sequence
 from copy import copy, deepcopy
-from functools import partial
+from functools import partial, singledispatchmethod
 from pathlib import Path
 from textwrap import dedent
-from typing import TYPE_CHECKING
+from typing import TYPE_CHECKING, cast
 
 import h5py
 import numpy as np
@@ -23,7 +23,7 @@ from scipy.sparse import issparse
 
 from .. import utils
 from .._settings import settings
-from ..compat import SpArray, _move_adj_mtx
+from ..compat import CSArray, _move_adj_mtx, old_positionals
 from ..logging import anndata_logger as logger
 from ..utils import (
     axis_len,
@@ -38,39 +38,23 @@ from .index import _normalize_indices, _subset, get_vector
 from .raw import Raw
 from .sparse_dataset import BaseCompressedSparseDataset
 from .storage import coerce_array
-from .views import (
-    DataFrameView,
-    DictView,
-    _resolve_idxs,
-)
+from .views import DictView, _resolve_idxs, as_view
+from .xarray import Dataset2D
 
 if TYPE_CHECKING:
     from collections.abc import Iterable
     from os import PathLike
-    from typing import Any, Literal
+    from typing import Any, ClassVar, Literal
 
-    from ..typing import ArrayDataStructureType
+    from zarr.storage import StoreLike
+
+    from ..compat import Index1D, XDataset
+    from ..typing import XDataType
     from .aligned_mapping import AxisArraysView, LayersView, PairwiseArraysView
-    from .index import Index, Index1D
+    from .index import Index
 
 
-# for backwards compat
-def _find_corresponding_multicol_key(key, keys_multicol):
-    """Find the corresponding multicolumn key."""
-    for mk in keys_multicol:
-        if key.startswith(mk) and "of" in key:
-            return mk
-    return None
-
-
-# for backwards compat
-def _gen_keys_from_multicol_key(key_multicol, n_keys):
-    """Generates single-column keys from multicolumn key."""
-    keys = [f"{key_multicol}{i + 1:03}of{n_keys:03}" for i in range(n_keys)]
-    return keys
-
-
-class AnnData(metaclass=utils.DeprecationMixinMeta):
+class AnnData(metaclass=utils.DeprecationMixinMeta):  # noqa: PLW1641
     """\
     An annotated data matrix.
 
@@ -185,10 +169,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
     .. _scikit-learn: http://scikit-learn.org/
     """
 
-    _BACKED_ATTRS = ["X", "raw.X"]
+    _BACKED_ATTRS: ClassVar[list[str]] = ["X", "raw.X"]
 
     # backwards compat
-    _H5_ALIASES = dict(
+    _H5_ALIASES: ClassVar[dict[str, set[str]]] = dict(
         X={"X", "_X", "data", "_data"},
         obs={"obs", "_obs", "smp", "_smp"},
         var={"var", "_var"},
@@ -198,26 +182,39 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         layers={"layers", "_layers"},
     )
 
-    _H5_ALIASES_NAMES = dict(
+    _H5_ALIASES_NAMES: ClassVar[dict[str, set[str]]] = dict(
         obs={"obs_names", "smp_names", "row_names", "index"},
         var={"var_names", "col_names", "index"},
     )
 
-    def __init__(
+    _accessors: ClassVar[set[str]] = set()
+
+    @old_positionals(
+        "obsm",
+        "varm",
+        "layers",
+        "raw",
+        "dtype",
+        "shape",
+        "filename",
+        "filemode",
+        "asview",
+    )
+    def __init__(  # noqa: PLR0913
         self,
-        X: np.ndarray | sparse.spmatrix | pd.DataFrame | None = None,
+        X: XDataType | pd.DataFrame | None = None,
         obs: pd.DataFrame | Mapping[str, Iterable[Any]] | None = None,
         var: pd.DataFrame | Mapping[str, Iterable[Any]] | None = None,
         uns: Mapping[str, Any] | None = None,
+        *,
         obsm: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
         varm: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
-        layers: Mapping[str, np.ndarray | sparse.spmatrix] | None = None,
+        layers: Mapping[str, XDataType] | None = None,
         raw: Mapping[str, Any] | None = None,
         shape: tuple[int, int] | None = None,
-        filename: PathLike | None = None,
+        filename: PathLike[str] | str | None = None,
         filemode: Literal["r", "r+"] | None = None,
         asview: bool = False,
-        *,
         obsp: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
         varp: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
         oidx: Index1D | None = None,
@@ -229,7 +226,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 raise_value_error_if_multiindex_columns(attr, key)
         if asview:
             if not isinstance(X, AnnData):
-                raise ValueError("`X` has to be an AnnData object.")
+                msg = "`X` has to be an AnnData object."
+                raise ValueError(msg)
             self._init_as_view(X, oidx, vidx)
         else:
             self._init_as_actual(
@@ -250,19 +248,22 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     def _init_as_view(self, adata_ref: AnnData, oidx: Index, vidx: Index):
         if adata_ref.isbacked and adata_ref.is_view:
-            raise ValueError(
+            msg = (
                 "Currently, you cannot index repeatedly into a backed AnnData, "
                 "that is, you cannot make a view of a view."
             )
+            raise ValueError(msg)
         self._is_view = True
         if isinstance(oidx, int | np.integer):
             if not (-adata_ref.n_obs <= oidx < adata_ref.n_obs):
-                raise IndexError(f"Observation index `{oidx}` is out of range.")
+                msg = f"Observation index `{oidx}` is out of range."
+                raise IndexError(msg)
             oidx += adata_ref.n_obs * (oidx < 0)
             oidx = slice(oidx, oidx + 1, 1)
         if isinstance(vidx, int | np.integer):
             if not (-adata_ref.n_vars <= vidx < adata_ref.n_vars):
-                raise IndexError(f"Variable index `{vidx}` is out of range.")
+                msg = f"Variable index `{vidx}` is out of range."
+                raise IndexError(msg)
             vidx += adata_ref.n_vars * (vidx < 0)
             vidx = slice(vidx, vidx + 1, 1)
         if adata_ref.is_view:
@@ -284,8 +285,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             self._remove_unused_categories(adata_ref.obs, obs_sub, uns)
             self._remove_unused_categories(adata_ref.var, var_sub, uns)
         # set attributes
-        self._obs = DataFrameView(obs_sub, view_args=(self, "obs"))
-        self._var = DataFrameView(var_sub, view_args=(self, "var"))
+        self._obs = as_view(obs_sub, view_args=(self, "obs"))
+        self._var = as_view(var_sub, view_args=(self, "var"))
         self._uns = uns
 
         # set raw, easy, as it’s immutable anyways...
@@ -296,9 +297,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         else:
             self._raw = None
 
-    def _init_as_actual(
+    def _init_as_actual(  # noqa: PLR0912, PLR0913, PLR0915
         self,
         X=None,
+        *,
         obs=None,
         var=None,
         uns=None,
@@ -334,9 +336,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             # init from AnnData
             if isinstance(X, AnnData):
                 if any((obs, var, uns, obsm, varm, obsp, varp)):
-                    raise ValueError(
-                        "If `X` is a dict no further arguments must be provided."
-                    )
+                    msg = "If `X` is a dict no further arguments must be provided."
+                    raise ValueError(msg)
                 obs, var, uns, obsm, varm, obsp, varp, layers, raw = (
                     X.obs,
                     X.var,
@@ -371,12 +372,19 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         if X is not None:
             X = coerce_array(X, name="X")
             if shape is not None:
-                raise ValueError("`shape` needs to be `None` if `X` is not `None`.")
+                msg = "`shape` needs to be `None` if `X` is not `None`."
+                raise ValueError(msg)
             # data matrix and shape
             n_obs, n_vars = X.shape
             source = "X"
         else:
-            n_obs, n_vars = (None, None) if shape is None else shape
+            n_obs, n_vars = (
+                shape
+                if shape is not None
+                else _infer_shape(
+                    obs, var, obsm=obsm, varm=varm, layers=layers, obsp=obsp, varp=varp
+                )
+            )
             source = "shape"
 
         # annotations
@@ -393,7 +401,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             if isinstance(attr.index, pd.RangeIndex):
                 attr.index = idx
             elif not idx.equals(attr.index):
-                raise ValueError(f"Index of {attr_name} must match {x_name} of X.")
+                msg = f"Index of {attr_name} must match {x_name} of X."
+                raise ValueError(msg)
 
         # unstructured annotations
         self.uns = uns or OrderedDict()
@@ -411,9 +420,9 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             self._check_uniqueness()
 
         if self.filename:
-            assert not isinstance(
-                raw, Raw
-            ), "got raw from other adata but also filename?"
+            assert not isinstance(raw, Raw), (
+                "got raw from other adata but also filename?"
+            )
             if {"raw", "raw.X"} & set(self.file):
                 raw = dict(X=None, **raw)
         if not raw:
@@ -430,7 +439,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         self.layers = layers
         self.X = X
 
-    def __sizeof__(self, show_stratified=None, with_disk: bool = False) -> int:
+    @old_positionals("show_stratified", "with_disk")
+    def __sizeof__(
+        self, *, show_stratified: bool = False, with_disk: bool = False
+    ) -> int:
         def get_size(X) -> int:
             def cs_to_bytes(X) -> int:
                 return int(X.data.nbytes + X.indptr.nbytes + X.indices.nbytes)
@@ -463,10 +475,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         return sum(sizes.values())
 
     def _gen_repr(self, n_obs, n_vars) -> str:
-        if self.isbacked:
-            backed_at = f" backed at {str(self.filename)!r}"
-        else:
-            backed_at = ""
+        backed_at = f" backed at {str(self.filename)!r}" if self.isbacked else ""
         descr = f"AnnData object with n_obs × n_vars = {n_obs} × {n_vars}{backed_at}"
         for attr in [
             "obs",
@@ -491,10 +500,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     def __eq__(self, other):
         """Equality testing"""
-        raise NotImplementedError(
+        msg = (
             "Equality comparisons are not supported for AnnData objects, "
             "instead compare the desired attributes."
         )
+        raise NotImplementedError(msg)
 
     @property
     def shape(self) -> tuple[int, int]:
@@ -502,12 +512,12 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         return self.n_obs, self.n_vars
 
     @property
-    def X(self) -> ArrayDataStructureType | None:
+    def X(self) -> XDataType | None:
         """Data matrix of shape :attr:`n_obs` × :attr:`n_vars`."""
         return self.layers.get(None)
 
     @X.setter
-    def X(self, value: np.ndarray | sparse.spmatrix | SpArray | None):
+    def X(self, value: XDataType) -> None:
         self.layers[None] = value
 
     @X.deleter
@@ -575,7 +585,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         if value is None:
             del self.raw
         elif not isinstance(value, AnnData):
-            raise ValueError("Can only init raw attribute with an AnnData object.")
+            msg = "Can only init raw attribute with an AnnData object."
+            raise ValueError(msg)
         else:
             if self.is_view:
                 self._init_as_actual(self.copy())
@@ -597,9 +608,14 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         """Number of variables/features."""
         return len(self.var_names)
 
-    def _set_dim_df(self, value: pd.DataFrame, attr: Literal["obs", "var"]):
-        if not isinstance(value, pd.DataFrame):
-            raise ValueError(f"Can only assign pd.DataFrame to {attr}.")
+    def _set_dim_df(self, value: pd.DataFrame | XDataset, attr: Literal["obs", "var"]):
+        value = _gen_dataframe(
+            value,
+            [f"{attr}_names", f"{'row' if attr == 'obs' else 'col'}_names"],
+            source="shape",
+            attr=attr,
+            length=self.n_obs if attr == "obs" else self.n_vars,
+        )
         raise_value_error_if_multiindex_columns(value, attr)
         value_idx = self._prep_dim_index(value.index, attr)
         if self.is_view:
@@ -615,14 +631,14 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         If a pd.Index is passed, this will use a reference, otherwise a new index object is created.
         """
         if self.shape[attr == "var"] != len(value):
-            raise ValueError(
-                f"Length of passed value for {attr}_names is {len(value)}, but this AnnData has shape: {self.shape}"
-            )
+            msg = f"Length of passed value for {attr}_names is {len(value)}, but this AnnData has shape: {self.shape}"
+            raise ValueError(msg)
         if isinstance(value, pd.Index) and not isinstance(value.name, str | type(None)):
-            raise ValueError(
+            msg = (
                 f"AnnData expects .{attr}.index.name to be a string or None, "
                 f"but you passed a name of type {type(value.name).__name__!r}"
             )
+            raise ValueError(msg)
         else:
             value = pd.Index(value)
             if not isinstance(value.name, str | type(None)):
@@ -654,12 +670,12 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 v.index = value
 
     @property
-    def obs(self) -> pd.DataFrame:
+    def obs(self) -> pd.DataFrame | Dataset2D:
         """One-dimensional annotation of observations (`pd.DataFrame`)."""
         return self._obs
 
     @obs.setter
-    def obs(self, value: pd.DataFrame):
+    def obs(self, value: pd.DataFrame | XDataset):
         self._set_dim_df(value, "obs")
 
     @obs.deleter
@@ -677,12 +693,12 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         self._set_dim_index(names, "obs")
 
     @property
-    def var(self) -> pd.DataFrame:
+    def var(self) -> pd.DataFrame | Dataset2D:
         """One-dimensional annotation of variables/ features (`pd.DataFrame`)."""
         return self._var
 
     @var.setter
-    def var(self, value: pd.DataFrame):
+    def var(self, value: pd.DataFrame | XDataset):
         self._set_dim_df(value, "var")
 
     @var.deleter
@@ -710,9 +726,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
     @uns.setter
     def uns(self, value: MutableMapping):
         if not isinstance(value, MutableMapping):
-            raise ValueError(
-                "Only mutable mapping types (e.g. dict) are allowed for `.uns`."
-            )
+            msg = "Only mutable mapping types (e.g. dict) are allowed for `.uns`."
+            raise ValueError(msg)
         if isinstance(value, DictView):
             value = value.copy()
         if self.is_view:
@@ -789,7 +804,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     def uns_keys(self) -> list[str]:
         """List keys of unstructured annotation."""
-        return sorted(list(self._uns.keys()))
+        return sorted(self._uns.keys())
 
     @property
     def isbacked(self) -> bool:
@@ -814,7 +829,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         return self.file.filename
 
     @filename.setter
-    def filename(self, filename: PathLike | None):
+    def filename(self, filename: PathLike[str] | str | None):
         # convert early for later comparison
         filename = None if filename is None else Path(filename)
         # change from backing-mode back to full loading into memory
@@ -838,10 +853,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             else:
                 # change from memory to backing-mode
                 # write the content of self to disk
-                if self.raw is not None:
-                    as_dense = ("X", "raw/X")
-                else:
-                    as_dense = ("X",)
+                as_dense = ("X", "raw/X") if self.raw is not None else ("X",)
                 self.write(filename, as_dense=as_dense)
             # open new file for accessing
             self.file.open(filename, "r+")
@@ -876,8 +888,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         oidx, vidx = self._normalize_indices(index)
         return AnnData(self, oidx=oidx, vidx=vidx, asview=True)
 
+    @singledispatchmethod
+    @staticmethod
     def _remove_unused_categories(
-        self, df_full: pd.DataFrame, df_sub: pd.DataFrame, uns: dict[str, Any]
+        df_full: pd.DataFrame, df_sub: pd.DataFrame, uns: dict[str, Any]
     ):
         for k in df_full:
             if not isinstance(df_full[k].dtype, pd.CategoricalDtype):
@@ -919,7 +933,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
              New categories, the same number as the old categories.
         """
         if isinstance(categories, Mapping):
-            raise ValueError("Only list-like `categories` is supported.")
+            msg = "Only list-like `categories` is supported."
+            raise ValueError(msg)
         if key in self.obs:
             old_categories = self.obs[key].cat.categories.tolist()
             self.obs[key] = self.obs[key].cat.rename_categories(categories)
@@ -927,7 +942,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             old_categories = self.var[key].cat.categories.tolist()
             self.var[key] = self.var[key].cat.rename_categories(categories)
         else:
-            raise ValueError(f"{key} is neither in `.obs` nor in `.var`.")
+            msg = f"{key} is neither in `.obs` nor in `.var`."
+            raise ValueError(msg)
         # this is not a good solution
         # but depends on the scanpy conventions for storing the categorical key
         # as `groupby` in the `params` slot
@@ -975,6 +991,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 dont_modify = True
         else:
             dfs = [df]
+        del df
+
         for df in dfs:
             string_cols = [
                 key for key in df.columns if infer_dtype(df[key]) == "string"
@@ -992,11 +1010,12 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 if not np.array_equal(c.categories, sorted_categories):
                     c = c.reorder_categories(sorted_categories)
                 if dont_modify:
-                    raise RuntimeError(
+                    msg = (
                         "Please call `.strings_to_categoricals()` on full "
                         "AnnData, not on this view. You might encounter this"
                         "error message while copying or writing to disk."
                     )
+                    raise RuntimeError(msg)
                 df[key] = c
                 logger.info(f"... storing {key!r} as categorical")
 
@@ -1023,11 +1042,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         self._init_as_actual(adata_subset)
 
     # TODO: Update, possibly remove
-    def __setitem__(
-        self, index: Index, val: int | float | np.ndarray | sparse.spmatrix
-    ):
+    def __setitem__(self, index: Index, val: float | XDataType):
         if self.is_view:
-            raise ValueError("Object is view and cannot be accessed with `[]`.")
+            msg = "Object is view and cannot be accessed with `[]`."
+            raise ValueError(msg)
         obs, var = self._normalize_indices(index)
         if not self.isbacked:
             self.X[obs, var] = val
@@ -1048,15 +1066,13 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         """
         from anndata.compat import _safe_transpose
 
-        if not self.isbacked:
-            X = self.X
-        else:
-            X = self.file["X"]
+        X = self.X if not self.isbacked else self.file["X"]
         if self.is_view:
-            raise ValueError(
+            msg = (
                 "You’re trying to transpose a view of an `AnnData`, "
                 "which is currently not implemented. Call `.copy()` before transposing."
             )
+            raise ValueError(msg)
 
         return AnnData(
             X=_safe_transpose(X) if X is not None else None,
@@ -1096,29 +1112,32 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         if layer is not None:
             X = self.layers[layer]
         elif not self._has_X():
-            raise ValueError("X is None, cannot convert to dataframe.")
+            msg = "X is None, cannot convert to dataframe."
+            raise ValueError(msg)
         else:
             X = self.X
         if issparse(X):
             X = X.toarray()
         return pd.DataFrame(X, index=self.obs_names, columns=self.var_names)
 
-    def _get_X(self, use_raw=False, layer=None):
+    def _get_X(self, *, use_raw: bool = False, layer: str | None = None):
         """\
         Convenience method for getting expression values
         with common arguments and error handling.
         """
         is_layer = layer is not None
         if use_raw and is_layer:
-            raise ValueError(
+            msg = (
                 "Cannot use expression from both layer and raw. You provided:"
                 f"`use_raw={use_raw}` and `layer={layer}`"
             )
+            raise ValueError(msg)
         if is_layer:
             return self.layers[layer]
         elif use_raw:
             if self.raw is None:
-                raise ValueError("This AnnData doesn’t have a value in `.raw`.")
+                msg = "This AnnData doesn’t have a value in `.raw`."
+                raise ValueError(msg)
             return self.raw.X
         else:
             return self.X
@@ -1147,11 +1166,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             if "X" in self.layers:
                 pass
             else:
-                warnings.warn(
+                msg = (
                     "In a future version of AnnData, access to `.X` by passing"
-                    " `layer='X'` will be removed. Instead pass `layer=None`.",
-                    FutureWarning,
+                    " `layer='X'` will be removed. Instead pass `layer=None`."
                 )
+                warnings.warn(msg, FutureWarning, stacklevel=2)
                 layer = None
         return get_vector(self, k, "obs", "var", layer=layer)
 
@@ -1179,16 +1198,16 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             if "X" in self.layers:
                 pass
             else:
-                warnings.warn(
+                msg = (
                     "In a future version of AnnData, access to `.X` by passing "
-                    "`layer='X'` will be removed. Instead pass `layer=None`.",
-                    FutureWarning,
+                    "`layer='X'` will be removed. Instead pass `layer=None`."
                 )
+                warnings.warn(msg, FutureWarning, stacklevel=2)
                 layer = None
         return get_vector(self, k, "var", "obs", layer=layer)
 
-    @utils.deprecated("obs_vector")
-    def _get_obs_array(self, k, use_raw=False, layer=None):
+    @deprecated("obs_vector")
+    def _get_obs_array(self, k, use_raw=False, layer=None):  # noqa: FBT002
         """\
         Get an array from the layer (default layer='X') along the :attr:`obs`
         dimension by first looking up `obs.keys` and then :attr:`obs_names`.
@@ -1198,8 +1217,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         else:
             return self.raw.obs_vector(k)
 
-    @utils.deprecated("var_vector")
-    def _get_var_array(self, k, use_raw=False, layer=None):
+    @deprecated("var_vector")
+    def _get_var_array(self, k, use_raw=False, layer=None):  # noqa: FBT002
         """\
         Get an array from the layer (default layer='X') along the :attr:`var`
         dimension by first looking up `var.keys` and then :attr:`var_names`.
@@ -1211,12 +1230,14 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     def _mutated_copy(self, **kwargs):
         """Creating AnnData with attributes optionally specified via kwargs."""
-        if self.isbacked:
-            if "X" not in kwargs or (self.raw is not None and "raw" not in kwargs):
-                raise NotImplementedError(
-                    "This function does not currently handle backed objects "
-                    "internally, this should be dealt with before."
-                )
+        if self.isbacked and (
+            "X" not in kwargs or (self.raw is not None and "raw" not in kwargs)
+        ):
+            msg = (
+                "This function does not currently handle backed objects "
+                "internally, this should be dealt with before."
+            )
+            raise NotImplementedError(msg)
         new = {}
 
         for key in ["obs", "var", "obsm", "varm", "obsp", "varp", "layers"]:
@@ -1234,7 +1255,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             new["raw"] = self.raw.copy()
         return AnnData(**new)
 
-    def to_memory(self, copy=False) -> AnnData:
+    @old_positionals("copy")
+    def to_memory(self, *, copy: bool = False) -> AnnData:
         """Return a new AnnData object with all backed arrays loaded into memory.
 
         Params
@@ -1265,13 +1287,13 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         ]:
             attr = getattr(self, attr_name, None)
             if attr is not None:
-                new[attr_name] = to_memory(attr, copy)
+                new[attr_name] = to_memory(attr, copy=copy)
 
         if self.raw is not None:
             new["raw"] = {
-                "X": to_memory(self.raw.X, copy),
-                "var": to_memory(self.raw.var, copy),
-                "varm": to_memory(self.raw.varm, copy),
+                "X": to_memory(self.raw.X, copy=copy),
+                "var": to_memory(self.raw.var, copy=copy),
+                "varm": to_memory(self.raw.varm, copy=copy),
             }
 
         if self.isbacked:
@@ -1279,7 +1301,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
         return AnnData(**new)
 
-    def copy(self, filename: PathLike | None = None) -> AnnData:
+    def copy(self, filename: PathLike[str] | str | None = None) -> AnnData:
         """Full copy, optionally on disk."""
         if not self.isbacked:
             if self.is_view and self._has_X():
@@ -1296,19 +1318,19 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             from ..io import read_h5ad, write_h5ad
 
             if filename is None:
-                raise ValueError(
+                msg = (
                     "To copy an AnnData object in backed mode, "
                     "pass a filename: `.copy(filename='myfilename.h5ad')`. "
                     "To load the object into memory, use `.to_memory()`."
                 )
+                raise ValueError(msg)
             mode = self.file._filemode
             write_h5ad(filename, self)
             return read_h5ad(filename, backed=mode)
 
     @deprecated(
         "anndata.concat",
-        FutureWarning,
-        "See the tutorial for concat at: "
+        add_msg="See the tutorial for concat at: "
         "https://anndata.readthedocs.io/en/latest/concatenation.html",
         hide=False,
     )
@@ -1317,7 +1339,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         *adatas: AnnData,
         join: str = "inner",
         batch_key: str = "batch",
-        batch_categories: Sequence[Any] = None,
+        batch_categories: Sequence[Any] | None = None,
         uns_merge: str | None = None,
         index_unique: str | None = "-",
         fill_value=None,
@@ -1405,6 +1427,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         AnnData object with n_obs × n_vars = 6 × 2
             obs: 'anno1', 'anno2', 'batch'
             var: 'annoA-0', 'annoA-1', 'annoA-2', 'annoB-2'
+            layers: None
         >>> adata.X
         array([[2, 3],
                [5, 6],
@@ -1434,6 +1457,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         AnnData object with n_obs × n_vars = 6 × 4
             obs: 'anno1', 'anno2', 'batch'
             var: 'annoA-0', 'annoA-1', 'annoA-2', 'annoB-2'
+            layers: None
         >>> outer.var.T
                    a    b    c    d
         annoA-0  0.0  1.0  2.0  NaN
@@ -1536,13 +1560,14 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         from .merge import concat, merge_dataframes, merge_outer, merge_same
 
         if self.isbacked:
-            raise ValueError("Currently, concatenate only works in memory mode.")
+            msg = "Currently, concatenate only works in memory mode."
+            raise ValueError(msg)
 
         if len(adatas) == 0:
             return self.copy()
         elif len(adatas) == 1 and not isinstance(adatas[0], AnnData):
             adatas = adatas[0]  # backwards compatibility
-        all_adatas = (self,) + tuple(adatas)
+        all_adatas = (self, *adatas)
 
         out = concat(
             all_adatas,
@@ -1559,7 +1584,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         # Backwards compat (some of this could be more efficient)
         # obs used to always be an outer join
         sparse_class = sparse.csr_matrix
-        if any(isinstance(a.X, SpArray) for a in all_adatas):
+        if any(isinstance(a.X, CSArray) for a in all_adatas):
             sparse_class = sparse.csr_array
         out.obs = concat(
             [AnnData(sparse_class(a.shape), obs=a.obs) for a in all_adatas],
@@ -1610,37 +1635,36 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             utils.warn_names_duplicates("var")
 
     def __contains__(self, key: Any):
-        raise AttributeError(
-            "AnnData has no attribute __contains__, don’t check `in adata`."
-        )
+        msg = "AnnData has no attribute __contains__, don’t check `in adata`."
+        raise AttributeError(msg)
 
     def _check_dimensions(self, key=None):
-        if key is None:
-            key = {"obsm", "varm"}
-        else:
-            key = {key}
-        if "obsm" in key:
-            if (
-                not all([axis_len(o, 0) == self.n_obs for o in self.obsm.values()])
-                and len(self.obsm.dim_names) != self.n_obs
-            ):
-                raise ValueError(
-                    "Observations annot. `obsm` must have number of rows of `X`"
-                    f" ({self.n_obs}), but has {len(self.obsm)} rows."
-                )
-        if "varm" in key:
-            if (
-                not all([axis_len(v, 0) == self.n_vars for v in self.varm.values()])
-                and len(self.varm.dim_names) != self.n_vars
-            ):
-                raise ValueError(
-                    "Variables annot. `varm` must have number of columns of `X`"
-                    f" ({self.n_vars}), but has {len(self.varm)} rows."
-                )
+        key = {"obsm", "varm"} if key is None else {key}
+        if "obsm" in key and (
+            not all(axis_len(o, 0) == self.n_obs for o in self.obsm.values())
+            and len(self.obsm.dim_names) != self.n_obs
+        ):
+            msg = (
+                "Observations annot. `obsm` must have number of rows of `X`"
+                f" ({self.n_obs}), but has {len(self.obsm)} rows."
+            )
+            raise ValueError(msg)
+        if "varm" in key and (
+            not all(axis_len(v, 0) == self.n_vars for v in self.varm.values())
+            and len(self.varm.dim_names) != self.n_vars
+        ):
+            msg = (
+                "Variables annot. `varm` must have number of columns of `X`"
+                f" ({self.n_vars}), but has {len(self.varm)} rows."
+            )
+            raise ValueError(msg)
 
+    @old_positionals("compression", "compression_opts", "as_dense")
     def write_h5ad(
         self,
-        filename: PathLike | None = None,
+        filename: PathLike[str] | str | None = None,
+        *,
+        convert_strings_to_categoricals: bool = True,
         compression: Literal["gzip", "lzf"] | None = None,
         compression_opts: int | Any = None,
         as_dense: Sequence[str] = (),
@@ -1664,6 +1688,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         ----------
         filename
             Filename of data file. Defaults to backing file.
+        convert_strings_to_categoricals
+            Convert string columns to categorical.
         compression
             For [`lzf`, `gzip`], see the h5py :ref:`dataset_compression`.
 
@@ -1710,13 +1736,15 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         from ..io import write_h5ad
 
         if filename is None and not self.isbacked:
-            raise ValueError("Provide a filename!")
+            msg = "Provide a filename!"
+            raise ValueError(msg)
         if filename is None:
             filename = self.filename
 
         write_h5ad(
             Path(filename),
             self,
+            convert_strings_to_categoricals=convert_strings_to_categoricals,
             compression=compression,
             compression_opts=compression_opts,
             as_dense=as_dense,
@@ -1727,7 +1755,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
     write = write_h5ad  # a shortcut and backwards compat
 
-    def write_csvs(self, dirname: PathLike, skip_data: bool = True, sep: str = ","):
+    @old_positionals("skip_data", "sep")
+    def write_csvs(
+        self, dirname: PathLike[str] | str, *, skip_data: bool = True, sep: str = ","
+    ):
         """\
         Write annotation to `.csv` files.
 
@@ -1747,7 +1778,10 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
         write_csvs(dirname, self, skip_data=skip_data, sep=sep)
 
-    def write_loom(self, filename: PathLike, write_obsm_varm: bool = False):
+    @old_positionals("write_obsm_varm")
+    def write_loom(
+        self, filename: PathLike[str] | str, *, write_obsm_varm: bool = False
+    ):
         """\
         Write `.loom`-formatted hdf5 file.
 
@@ -1760,10 +1794,13 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
 
         write_loom(filename, self, write_obsm_varm=write_obsm_varm)
 
+    @old_positionals("chunks")
     def write_zarr(
         self,
-        store: MutableMapping | PathLike,
-        chunks: bool | int | tuple[int, ...] | None = None,
+        store: StoreLike,
+        *,
+        chunks: tuple[int, ...] | None = None,
+        convert_strings_to_categoricals: bool = True,
     ):
         """\
         Write a hierarchical Zarr array store.
@@ -1774,10 +1811,24 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
             The filename, a :class:`~typing.MutableMapping`, or a Zarr storage class.
         chunks
             Chunk shape.
+        convert_strings_to_categoricals
+            Convert string columns to categorical.
         """
         from ..io import write_zarr
 
-        write_zarr(store, self, chunks=chunks)
+        # TODO: What is a bool for chunks supposed to do?
+        if isinstance(chunks, bool):
+            msg = (
+                "Passing `write_zarr(adata, chunks=True)` is no longer supported. "
+                "Please pass `write_zarr(adata)` instead."
+            )
+            raise ValueError(msg)
+        write_zarr(
+            store,
+            self,
+            chunks=chunks,
+            convert_strings_to_categoricals=convert_strings_to_categoricals,
+        )
 
     def chunked_X(self, chunk_size: int | None = None):
         """\
@@ -1800,9 +1851,11 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         if start < n:
             yield (self.X[start:n], start, n)
 
+    @old_positionals("replace")
     def chunk_X(
         self,
         select: int | Sequence[int] | np.ndarray = 1000,
+        *,
         replace: bool = True,
     ):
         """\
@@ -1828,7 +1881,8 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
         elif isinstance(select, np.ndarray | Sequence):
             choice = np.asarray(select)
         else:
-            raise ValueError("select should be int or array")
+            msg = "select should be int or array"
+            raise ValueError(msg)
 
         reverse = None
         if self.isbacked:
@@ -1860,7 +1914,7 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
     # --------------------------------------------------------------------------
 
     @property
-    @utils.deprecated("is_view")
+    @deprecated("is_view")
     def isview(self):
         return self.is_view
 
@@ -1883,10 +1937,62 @@ class AnnData(metaclass=utils.DeprecationMixinMeta):
                 m_attr[key] = self._get_and_delete_multicol_field(axis, key)
 
     def _get_and_delete_multicol_field(self, a, key_multicol):
-        keys = []
-        for k in getattr(self, a).columns:
-            if k.startswith(key_multicol):
-                keys.append(k)
+        keys = [k for k in getattr(self, a).columns if k.startswith(key_multicol)]
         values = getattr(self, a)[keys].values
         getattr(self, a).drop(keys, axis=1, inplace=True)
         return values
+
+
+@AnnData._remove_unused_categories.register(Dataset2D)
+@staticmethod
+def _remove_unused_categories_xr(
+    df_full: Dataset2D, df_sub: Dataset2D, uns: dict[str, Any]
+):
+    pass  # this is handled automatically by the categorical arrays themselves i.e., they dedup upon access.
+
+
+def _check_2d_shape(X):
+    """\
+    Check shape of array or sparse matrix.
+
+    Assure that X is always 2D: Unlike numpy we always deal with 2D arrays.
+    """
+    if X.dtype.names is None and len(X.shape) != 2:
+        msg = f"X needs to be 2-dimensional, not {len(X.shape)}-dimensional."
+        raise ValueError(msg)
+
+
+def _infer_shape_for_axis(
+    xxx: pd.DataFrame | Mapping[str, Iterable[Any]] | None,
+    xxxm: np.ndarray | Mapping[str, Sequence[Any]] | None,
+    layers: Mapping[str, np.ndarray | sparse.spmatrix] | None,
+    xxxp: np.ndarray | Mapping[str, Sequence[Any]] | None,
+    axis: Literal[0, 1],
+) -> int | None:
+    for elem in [xxx, xxxm, xxxp]:
+        if elem is not None and hasattr(elem, "shape"):
+            return elem.shape[0]
+    for elem, id in zip([layers, xxxm, xxxp], ["layers", "xxxm", "xxxp"], strict=True):
+        if elem is not None:
+            elem = cast("Mapping", elem)
+            for sub_elem in elem.values():
+                if hasattr(sub_elem, "shape"):
+                    size = cast("int", sub_elem.shape[axis if id == "layers" else 0])
+                    return size
+    return None
+
+
+def _infer_shape(
+    obs: pd.DataFrame | Mapping[str, Iterable[Any]] | None = None,
+    var: pd.DataFrame | Mapping[str, Iterable[Any]] | None = None,
+    *,
+    obsm: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
+    varm: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
+    layers: Mapping[str, np.ndarray | sparse.spmatrix] | None = None,
+    obsp: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
+    varp: np.ndarray | Mapping[str, Sequence[Any]] | None = None,
+):
+    return (
+        _infer_shape_for_axis(obs, obsm, layers, obsp, 0),
+        _infer_shape_for_axis(var, varm, layers, varp, 1),
+    )
