@@ -8,9 +8,10 @@ from typing import TYPE_CHECKING
 import h5py
 import numpy as np
 import pandas as pd
-from scipy.sparse import issparse, spmatrix
+from scipy.sparse import issparse
 
-from ..compat import AwkArray, DaskArray, SpArray
+from ..compat import AwkArray, CSArray, CSMatrix, DaskArray, XDataArray
+from .xarray import Dataset2D
 
 if TYPE_CHECKING:
     from ..compat import Index, Index1D
@@ -26,21 +27,15 @@ def _normalize_indices(
     if isinstance(index, pd.Series):
         index: Index = index.values
     if isinstance(index, tuple):
-        if len(index) > 2:
-            raise ValueError("AnnData can only be sliced in rows and columns.")
-        # deal with pd.Series
         # TODO: The series should probably be aligned first
-        if isinstance(index[1], pd.Series):
-            index = index[0], index[1].values
-        if isinstance(index[0], pd.Series):
-            index = index[0].values, index[1]
+        index = tuple(i.values if isinstance(i, pd.Series) else i for i in index)
     ax0, ax1 = unpack_index(index)
     ax0 = _normalize_index(ax0, names0)
     ax1 = _normalize_index(ax1, names1)
     return ax0, ax1
 
 
-def _normalize_index(
+def _normalize_index(  # noqa: PLR0911, PLR0912
     indexer: slice
     | np.integer
     | int
@@ -50,10 +45,10 @@ def _normalize_index(
     | pd.Index,
     index: pd.Index,
 ) -> slice | int | np.ndarray:  # ndarray of int or bool
-    if not isinstance(index, pd.RangeIndex):
-        msg = "Don’t call _normalize_index with non-categorical/string names"
-        assert index.dtype != float, msg
-        assert index.dtype != int, msg
+    # TODO: why is this here? All tests pass without it and it seems at the minimum not strict enough.
+    if not isinstance(index, pd.RangeIndex) and index.dtype in (np.float64, np.int64):
+        msg = f"Don’t call _normalize_index with non-categorical/string names and non-range index {index}"
+        raise TypeError(msg)
 
     # the following is insanely slow for sequences,
     # we replaced it using pandas below
@@ -75,40 +70,53 @@ def _normalize_index(
     elif isinstance(indexer, str):
         return index.get_loc(indexer)  # int
     elif isinstance(
-        indexer, Sequence | np.ndarray | pd.Index | spmatrix | np.matrix | SpArray
+        indexer, Sequence | np.ndarray | pd.Index | CSMatrix | np.matrix | CSArray
     ):
         if hasattr(indexer, "shape") and (
             (indexer.shape == (index.shape[0], 1))
             or (indexer.shape == (1, index.shape[0]))
         ):
-            if isinstance(indexer, spmatrix | SpArray):
+            if isinstance(indexer, CSMatrix | CSArray):
                 indexer = indexer.toarray()
             indexer = np.ravel(indexer)
         if not isinstance(indexer, np.ndarray | pd.Index):
             indexer = np.array(indexer)
             if len(indexer) == 0:
                 indexer = indexer.astype(int)
+        if isinstance(indexer, np.ndarray) and np.issubdtype(
+            indexer.dtype, np.floating
+        ):
+            indexer_int = indexer.astype(int)
+            if np.all((indexer - indexer_int) != 0):
+                msg = f"Indexer {indexer!r} has floating point values."
+                raise IndexError(msg)
         if issubclass(indexer.dtype.type, np.integer | np.floating):
             return indexer  # Might not work for range indexes
         elif issubclass(indexer.dtype.type, np.bool_):
             if indexer.shape != index.shape:
-                raise IndexError(
+                msg = (
                     f"Boolean index does not match AnnData’s shape along this "
                     f"dimension. Boolean index has shape {indexer.shape} while "
                     f"AnnData index has shape {index.shape}."
                 )
+                raise IndexError(msg)
             return indexer
         else:  # indexer should be string array
             positions = index.get_indexer(indexer)
             if np.any(positions < 0):
                 not_found = indexer[positions < 0]
-                raise KeyError(
+                msg = (
                     f"Values {list(not_found)}, from {list(indexer)}, "
                     "are not valid obs/ var names or indices."
                 )
+                raise KeyError(msg)
             return positions  # np.ndarray[int]
-    else:
-        raise IndexError(f"Unknown indexer {indexer!r} of type {type(indexer)}")
+    elif isinstance(indexer, XDataArray):
+        if isinstance(indexer.data, DaskArray):
+            return indexer.data.compute()
+        return indexer.data
+    msg = f"Unknown indexer {indexer!r} of type {type(indexer)}"
+    raise IndexError()
 
 
 def _fix_slice_bounds(s: slice, length: int) -> slice:
@@ -132,13 +140,31 @@ def _fix_slice_bounds(s: slice, length: int) -> slice:
 
 def unpack_index(index: Index) -> tuple[Index1D, Index1D]:
     if not isinstance(index, tuple):
+        if index is Ellipsis:
+            index = slice(None)
         return index, slice(None)
-    elif len(index) == 2:
+    num_ellipsis = sum(i is Ellipsis for i in index)
+    if num_ellipsis > 1:
+        msg = "an index can only have a single ellipsis ('...')"
+        raise IndexError(msg)
+    # If index has Ellipsis, filter it out (and if not, error)
+    if len(index) > 2:
+        if not num_ellipsis:
+            msg = "Received a length 3 index without an ellipsis"
+            raise IndexError(msg)
+        index = tuple(i for i in index if i is not Ellipsis)
         return index
-    elif len(index) == 1:
-        return index[0], slice(None)
-    else:
-        raise IndexError("invalid number of indices")
+    # If index has Ellipsis, replace it with slice
+    if len(index) == 2:
+        index = tuple(slice(None) if i is Ellipsis else i for i in index)
+        return index
+    if len(index) == 1:
+        index = index[0]
+        if index is Ellipsis:
+            index = slice(None)
+        return index, slice(None)
+    msg = "invalid number of indices"
+    raise IndexError(msg)
 
 
 @singledispatch
@@ -159,9 +185,9 @@ def _subset_dask(a: DaskArray, subset_idx: Index):
     return a[subset_idx]
 
 
-@_subset.register(spmatrix)
-@_subset.register(SpArray)
-def _subset_sparse(a: spmatrix | SpArray, subset_idx: Index):
+@_subset.register(CSMatrix)
+@_subset.register(CSArray)
+def _subset_sparse(a: CSMatrix | CSArray, subset_idx: Index):
     # Correcting for indexing behaviour of sparse.spmatrix
     if len(subset_idx) > 1 and all(isinstance(x, Iterable) for x in subset_idx):
         first_idx = subset_idx[0]
@@ -172,7 +198,8 @@ def _subset_sparse(a: spmatrix | SpArray, subset_idx: Index):
 
 
 @_subset.register(pd.DataFrame)
-def _subset_df(df: pd.DataFrame, subset_idx: Index):
+@_subset.register(Dataset2D)
+def _subset_df(df: pd.DataFrame | Dataset2D, subset_idx: Index):
     return df.iloc[subset_idx]
 
 
@@ -217,13 +244,11 @@ def get_vector(adata, k, coldim, idxdim, layer=None):
     in_idx = k in idx
 
     if (in_col + in_idx) == 2:
-        raise ValueError(
-            f"Key {k} could be found in both .{idxdim}_names and .{coldim}.columns"
-        )
+        msg = f"Key {k} could be found in both .{idxdim}_names and .{coldim}.columns"
+        raise ValueError(msg)
     elif (in_col + in_idx) == 0:
-        raise KeyError(
-            f"Could not find key {k} in .{idxdim}_names or .{coldim}.columns."
-        )
+        msg = f"Could not find key {k} in .{idxdim}_names or .{coldim}.columns."
+        raise KeyError(msg)
     elif in_col:
         return getattr(adata, coldim)[k].values
     elif in_idx:
