@@ -9,9 +9,9 @@ def __getattr__(key: str):
     from .. import io
 
     attr = getattr(io, key)
-    warnings.warn(
+    msg = (
         f"Importing {key} from `anndata._io` is deprecated. "
-        "Please use anndata.io instead.",
-        FutureWarning,
+        "Please use anndata.io instead."
     )
+    warnings.warn(msg, FutureWarning, stacklevel=2)
     return attr
