@@ -15,16 +15,18 @@ import pandas as pd
 from scipy import sparse
 
 from .. import AnnData
-from ..compat import _deprecate_positional_args
+from ..compat import old_positionals
 from .utils import is_float
 
 if TYPE_CHECKING:
     from collections.abc import Generator, Iterable, Iterator, Mapping
 
 
+@old_positionals("first_column_names", "dtype")
 def read_csv(
-    filename: PathLike | Iterator[str],
+    filename: PathLike[str] | str | Iterator[str],
     delimiter: str | None = ",",
+    *,
     first_column_names: bool | None = None,
     dtype: str = "float32",
 ) -> AnnData:
@@ -49,7 +51,9 @@ def read_csv(
     return read_text(filename, delimiter, first_column_names, dtype)
 
 
-def read_excel(filename: PathLike, sheet: str | int, dtype: str = "float32") -> AnnData:
+def read_excel(
+    filename: PathLike[str] | str, sheet: str | int, dtype: str = "float32"
+) -> AnnData:
     """\
     Read `.xlsx` (Excel) file.
 
@@ -73,7 +77,7 @@ def read_excel(filename: PathLike, sheet: str | int, dtype: str = "float32") -> 
     return AnnData(X, row, col)
 
 
-def read_umi_tools(filename: PathLike, dtype=None) -> AnnData:
+def read_umi_tools(filename: PathLike[str] | str, dtype=None) -> AnnData:
     """\
     Read a gzipped condensed count matrix from umi_tools.
 
@@ -96,7 +100,7 @@ def read_umi_tools(filename: PathLike, dtype=None) -> AnnData:
     return AnnData(X=X, obs=obs, var=var)
 
 
-def read_hdf(filename: PathLike, key: str) -> AnnData:
+def read_hdf(filename: PathLike[str] | str, key: str) -> AnnData:
     """\
     Read `.h5` (hdf5) file.
 
@@ -112,12 +116,13 @@ def read_hdf(filename: PathLike, key: str) -> AnnData:
     with h5py.File(filename, "r") as f:
         # the following is necessary in Python 3, because only
         # a view and not a list is returned
-        keys = [k for k in f.keys()]
+        keys = list(f)
         if key == "":
-            raise ValueError(
+            msg = (
                 f"The file {filename} stores the following sheets:\n{keys}\n"
                 f"Call read/read_hdf5 with one of them."
             )
+            raise ValueError(msg)
         # read array
         X = f[key][()]
         # try to find row and column names
@@ -149,9 +154,20 @@ def _fmt_loom_axis_attrs(
     return axis_df, axis_mapping
 
 
-@_deprecate_positional_args(version="0.9")
-def read_loom(
-    filename: PathLike,
+@old_positionals(
+    "sparse",
+    "cleanup",
+    "X_name",
+    "obs_names",
+    "obsm_names",
+    "var_names",
+    "varm_names",
+    "dtype",
+    "obsm_mapping",
+    "varm_mapping",
+)
+def read_loom(  # noqa: PLR0912, PLR0913
+    filename: PathLike[str] | str,
     *,
     sparse: bool = True,
     cleanup: bool = False,
@@ -213,35 +229,43 @@ def read_loom(
     """
     # Deprecations
     if obsm_names is not None:
-        warn(
+        msg = (
             "Argument obsm_names has been deprecated in favour of `obsm_mapping`. "
-            "In 0.9 this will be an error.",
-            FutureWarning,
+            "In 0.9 this will be an error."
         )
+        warn(msg, FutureWarning, stacklevel=2)
         if obsm_mapping != {}:
-            raise ValueError(
+            msg = (
                 "Received values for both `obsm_names` and `obsm_mapping`. This is "
                 "ambiguous, only pass `obsm_mapping`."
             )
+            raise ValueError(msg)
         obsm_mapping = obsm_names
     if varm_names is not None:
-        warn(
+        msg = (
             "Argument varm_names has been deprecated in favour of `varm_mapping`. "
-            "In 0.9 this will be an error.",
-            FutureWarning,
+            "In 0.9 this will be an error."
         )
+        warn(msg, FutureWarning, stacklevel=2)
         if varm_mapping != {}:
-            raise ValueError(
+            msg = (
                 "Received values for both `varm_names` and `varm_mapping`. This is "
                 "ambiguous, only pass `varm_mapping`."
             )
+            raise ValueError(msg)
         varm_mapping = varm_names
 
     filename = fspath(filename)  # allow passing pathlib.Path objects
     from loompy import connect
 
+    if TYPE_CHECKING:
+        from loompy import LoomConnection
+
+    lc: LoomConnection
     with connect(filename, "r", **kwargs) as lc:
-        if X_name not in lc.layers.keys():
+        assert lc.layers is not None
+
+        if X_name not in lc.layers:
             X_name = ""
         X = lc.layers[X_name].sparse().T.tocsr() if sparse else lc.layers[X_name][()].T
         X = X.astype(dtype, copy=False)
@@ -251,13 +275,9 @@ def read_loom(
             layers["matrix"] = (
                 lc.layers[""].sparse().T.tocsr() if sparse else lc.layers[""][()].T
             )
-        for key in lc.layers.keys():
+        for key, layer in lc.layers.items():
             if key != "":
-                layers[key] = (
-                    lc.layers[key].sparse().T.tocsr()
-                    if sparse
-                    else lc.layers[key][()].T
-                )
+                layers[key] = layer.sparse().T.tocsr() if sparse else layer[()].T
 
         # TODO: Figure out the singleton obs elements
         obs, obsm = _fmt_loom_axis_attrs(dict(lc.col_attrs), obs_names, obsm_mapping)
@@ -292,7 +312,7 @@ def read_loom(
     return adata
 
 
-def read_mtx(filename: PathLike, dtype: str = "float32") -> AnnData:
+def read_mtx(filename: PathLike[str] | str, dtype: str = "float32") -> AnnData:
     """\
     Read `.mtx` file.
 
@@ -313,9 +333,11 @@ def read_mtx(filename: PathLike, dtype: str = "float32") -> AnnData:
     return AnnData(X)
 
 
+@old_positionals("first_column_names", "dtype")
 def read_text(
-    filename: PathLike | Iterator[str],
+    filename: PathLike[str] | str | Iterator[str],
     delimiter: str | None = None,
+    *,
     first_column_names: bool | None = None,
     dtype: str = "float32",
 ) -> AnnData:
@@ -360,10 +382,10 @@ def _iter_lines(file_like: Iterable[str]) -> Generator[str, None, None]:
             yield line
 
 
-def _read_text(
+def _read_text(  # noqa: PLR0912, PLR0915
     f: Iterator[str],
     delimiter: str | None,
-    first_column_names: bool | None,
+    first_column_names: bool | None,  # noqa: FBT001
     dtype: str,
 ) -> AnnData:
     comments = []
@@ -379,19 +401,19 @@ def _read_text(
                 comments.append(comment)
         else:
             if delimiter is not None and delimiter not in line:
-                raise ValueError(f"Did not find delimiter {delimiter!r} in first line.")
+                msg = f"Did not find delimiter {delimiter!r} in first line."
+                raise ValueError(msg)
             line_list = line.split(delimiter)
             # the first column might be row names, so check the last
             if not is_float(line_list[-1]):
                 col_names = line_list
                 # logg.msg("    assuming first line in file stores column names", v=4)
+            elif not is_float(line_list[0]) or first_column_names:
+                first_column_names = True
+                row_names.append(line_list[0])
+                data.append(np.array(line_list[1:], dtype=dtype))
             else:
-                if not is_float(line_list[0]) or first_column_names:
-                    first_column_names = True
-                    row_names.append(line_list[0])
-                    data.append(np.array(line_list[1:], dtype=dtype))
-                else:
-                    data.append(np.array(line_list, dtype=dtype))
+                data.append(np.array(line_list, dtype=dtype))
             break
     if not col_names:
         # try reading col_names from the last comment line
@@ -441,10 +463,11 @@ def _read_text(
     #   in the end, to separate row_names from float data, slicing takes
     #   a lot of memory and CPU time
     if data[0].size != data[-1].size:
-        raise ValueError(
+        msg = (
             f"Length of first line ({data[0].size}) is different "
             f"from length of last line ({data[-1].size})."
         )
+        raise ValueError(msg)
     data = np.array(data, dtype=dtype)
     # logg.msg("    constructed array from list of list", t=True, v=4)
     # transform row_names
