@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import inspect
 import warnings
 from collections.abc import Mapping
 from dataclasses import dataclass
@@ -8,8 +9,8 @@ from types import MappingProxyType
 from typing import TYPE_CHECKING, Generic, TypeVar
 
 from anndata._io.utils import report_read_key_on_error, report_write_key_on_error
-from anndata._types import Read, ReadDask, _ReadDaskInternal, _ReadInternal
-from anndata.compat import DaskArray, _read_attr
+from anndata._types import Read, ReadLazy, _ReadInternal, _ReadLazyInternal
+from anndata.compat import DaskArray, ZarrGroup, _read_attr, is_zarr_v2
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Generator, Iterable
@@ -23,10 +24,14 @@ if TYPE_CHECKING:
         WriteCallback,
         _WriteInternal,
     )
+    from anndata.experimental.backed._lazy_arrays import CategoricalArray, MaskedArray
     from anndata.typing import RWAble
+
+    from ..._core.xarray import Dataset2D
 
     T = TypeVar("T")
     W = TypeVar("W", bound=_WriteInternal)
+    LazyDataStructures = DaskArray | Dataset2D | CategoricalArray | MaskedArray
 
 
 # TODO: This probably should be replaced by a hashable Mapping due to conversion b/w "_" and "-"
@@ -78,8 +83,8 @@ def write_spec(spec: IOSpec):
     return decorator
 
 
-_R = TypeVar("_R", _ReadInternal, _ReadDaskInternal)
-R = TypeVar("R", Read, ReadDask)
+_R = TypeVar("_R", _ReadInternal, _ReadLazyInternal)
+R = TypeVar("R", Read, ReadLazy)
 
 
 class IORegistry(Generic[_R, R]):
@@ -105,10 +110,11 @@ class IORegistry(Generic[_R, R]):
         if src_type in self.write_specs and (spec != self.write_specs[src_type]):
             # First check for consistency
             current_spec = self.write_specs[src_type]
-            raise TypeError(
+            msg = (
                 "Cannot overwrite IO specifications. Attempted to overwrite encoding "
                 f"for {src_type} from {current_spec} to {spec}"
             )
+            raise TypeError(msg)
         else:
             self.write_specs[src_type] = spec
 
@@ -130,7 +136,6 @@ class IORegistry(Generic[_R, R]):
 
         if dest_type is h5py.File:
             dest_type = h5py.Group
-
         if (dest_type, src_type, modifiers) not in self.write:
             raise IORegistryError._from_write_parts(dest_type, src_type, modifiers)
         internal = self.write[(dest_type, src_type, modifiers)]
@@ -168,7 +173,7 @@ class IORegistry(Generic[_R, R]):
         reader: Reader,
     ) -> R:
         if (src_type, spec, modifiers) not in self.read:
-            raise IORegistryError._from_read_parts("read", self.read, src_type, spec)
+            raise IORegistryError._from_read_parts("read", self.read, src_type, spec)  # noqa: EM101
         internal = self.read[(src_type, spec, modifiers)]
         return partial(internal, _reader=reader)
 
@@ -197,28 +202,29 @@ class IORegistry(Generic[_R, R]):
     ):
         if (src_type, spec, modifiers) in self.read_partial:
             return self.read_partial[(src_type, spec, modifiers)]
-        else:
-            raise IORegistryError._from_read_parts(
-                "read_partial", self.read_partial, src_type, spec
-            )
+        name = "read_partial"
+        raise IORegistryError._from_read_parts(name, self.read_partial, src_type, spec)
 
     def get_spec(self, elem: Any) -> IOSpec:
         if isinstance(elem, DaskArray):
             if (typ_meta := (DaskArray, type(elem._meta))) in self.write_specs:
                 return self.write_specs[typ_meta]
-        elif hasattr(elem, "dtype"):
-            if (typ_kind := (type(elem), elem.dtype.kind)) in self.write_specs:
-                return self.write_specs[typ_kind]
+        elif (
+            hasattr(elem, "dtype")
+            and (typ_kind := (type(elem), elem.dtype.kind)) in self.write_specs
+        ):
+            return self.write_specs[typ_kind]
         return self.write_specs[type(elem)]
 
 
 _REGISTRY: IORegistry[_ReadInternal, Read] = IORegistry()
-_LAZY_REGISTRY: IORegistry[_ReadDaskInternal, ReadDask] = IORegistry()
+_LAZY_REGISTRY: IORegistry[_ReadLazyInternal, ReadLazy] = IORegistry()
 
 
 @singledispatch
 def proc_spec(spec) -> IOSpec:
-    raise NotImplementedError(f"proc_spec not defined for type: {type(spec)}.")
+    msg = f"proc_spec not defined for type: {type(spec)}."
+    raise NotImplementedError(msg)
 
 
 @proc_spec.register(IOSpec)
@@ -282,24 +288,35 @@ class Reader:
         return self.callback(read_func, elem.name, elem, iospec=iospec)
 
 
-class DaskReader(Reader):
+class LazyReader(Reader):
     @report_read_key_on_error
     def read_elem(
         self,
         elem: StorageType,
         modifiers: frozenset[str] = frozenset(),
         chunks: tuple[int, ...] | None = None,
-    ) -> DaskArray:
+        **kwargs,
+    ) -> LazyDataStructures:
         """Read a dask element from a store. See exported function for more details."""
 
         iospec = get_spec(elem)
-        read_func: ReadDask = self.registry.get_read(
+        read_func: ReadLazy = self.registry.get_read(
             type(elem), iospec, modifiers, reader=self
         )
         if self.callback is not None:
             msg = "Dask reading does not use a callback. Ignoring callback."
             warnings.warn(msg, stacklevel=2)
-        return read_func(elem, chunks=chunks)
+        read_params = inspect.signature(read_func).parameters
+        for kwarg in kwargs:
+            if kwarg not in read_params:
+                msg = (
+                    f"Keyword argument {kwarg} passed to read_elem_lazy are not supported by the "
+                    "registered read function."
+                )
+                raise ValueError(msg)
+        if "chunks" in read_params:
+            kwargs["chunks"] = chunks
+        return read_func(elem, **kwargs)
 
 
 class Writer:
@@ -332,20 +349,40 @@ class Writer:
 
         import h5py
 
+        # we allow stores to have a prefix like /uns which are then written to with keys like /uns/foo
+        if "/" in k.split(store.name)[-1][1:]:
+            msg = "Forward slashes are not allowed in keys."
+            raise ValueError(msg)
+
         if isinstance(store, h5py.File):
             store = store["/"]
 
         dest_type = type(store)
 
-        if elem is None:
-            return lambda *_, **__: None
-
         # Normalize k to absolute path
-        if not PurePosixPath(k).is_absolute():
+        if (
+            is_zarr_v2_store := (
+                (is_zarr_store := isinstance(store, ZarrGroup)) and is_zarr_v2()
+            )
+        ) or (isinstance(store, h5py.Group) and not PurePosixPath(k).is_absolute()):
             k = str(PurePosixPath(store.name) / k)
+        is_consolidated = False
+        if is_zarr_v2_store:
+            from zarr.storage import ConsolidatedMetadataStore
 
+            is_consolidated = isinstance(store.store, ConsolidatedMetadataStore)
+        elif is_zarr_store:
+            is_consolidated = store.metadata.consolidated_metadata is not None
+        if is_consolidated:
+            msg = "Cannot overwrite/edit a store with consolidated metadata"
+            raise ValueError(msg)
         if k == "/":
-            store.clear()
+            if isinstance(store, ZarrGroup) and not is_zarr_v2():
+                from zarr.core.sync import sync
+
+                sync(store.store.clear())
+            else:
+                store.clear()
         elif k in store:
             del store[k]
 
@@ -378,9 +415,9 @@ def read_elem(elem: StorageType) -> RWAble:
     return Reader(_REGISTRY).read_elem(elem)
 
 
-def read_elem_as_dask(
-    elem: StorageType, chunks: tuple[int, ...] | None = None
-) -> DaskArray:
+def read_elem_lazy(
+    elem: StorageType, chunks: tuple[int, ...] | None = None, **kwargs
+) -> LazyDataStructures:
     """
     Read an element from a store lazily.
 
@@ -392,16 +429,17 @@ def read_elem_as_dask(
     ----------
     elem
         The stored element.
-    chunks, optional
+    chunks
        length `n`, the same `n` as the size of the underlying array.
        Note that the minor axis dimension must match the shape for sparse.
        Defaults to `(1000, adata.shape[1])` for CSR sparse,
        `(adata.shape[0], 1000)` for CSC sparse,
        and the on-disk chunking otherwise for dense.
+       Can use `-1` or `None` to indicate use of the size of the corresponding dimension.
 
     Returns
     -------
-        DaskArray
+        A "lazy" elem
 
     Examples
     --------
@@ -423,18 +461,16 @@ def read_elem_as_dask(
     Reading a sparse matrix from a zarr store lazily, with custom chunk size and default:
 
     >>> g = zarr.open(zarr_path)
-    >>> adata.X = ad.experimental.read_elem_as_dask(g["X"])
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"])
     >>> adata.X
     dask.array<make_dask_chunk, shape=(2700, 32738), dtype=float32, chunksize=(1000, 32738), chunktype=scipy.csr_matrix>
-    >>> adata.X = ad.experimental.read_elem_as_dask(
-    ...     g["X"], chunks=(500, adata.shape[1])
-    ... )
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"], chunks=(500, adata.shape[1]))
     >>> adata.X
     dask.array<make_dask_chunk, shape=(2700, 32738), dtype=float32, chunksize=(500, 32738), chunktype=scipy.csr_matrix>
 
     Reading a dense matrix from a zarr store lazily:
 
-    >>> adata.layers["dense"] = ad.experimental.read_elem_as_dask(g["layers/dense"])
+    >>> adata.layers["dense"] = ad.experimental.read_elem_lazy(g["layers/dense"])
     >>> adata.layers["dense"]
     dask.array<from-zarr, shape=(2700, 32738), dtype=float32, chunksize=(169, 2047), chunktype=numpy.ndarray>
 
@@ -447,12 +483,15 @@ def read_elem_as_dask(
     ...     obsm=ad.io.read_elem(g["obsm"]),
     ...     varm=ad.io.read_elem(g["varm"]),
     ... )
-    >>> adata.X = ad.experimental.read_elem_as_dask(
-    ...     g["X"], chunks=(500, adata.shape[1])
-    ... )
-    >>> adata.layers["dense"] = ad.experimental.read_elem_as_dask(g["layers/dense"])
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"], chunks=(500, adata.shape[1]))
+    >>> adata.layers["dense"] = ad.experimental.read_elem_lazy(g["layers/dense"])
+
+    We also support using -1 and None as a chunk size to signify the reading the whole axis:
+
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"], chunks=(500, -1))
+    >>> adata.X = ad.experimental.read_elem_lazy(g["X"], chunks=(500, None))
     """
-    return DaskReader(_LAZY_REGISTRY).read_elem(elem, chunks=chunks)
+    return LazyReader(_LAZY_REGISTRY).read_elem(elem, chunks=chunks, **kwargs)
 
 
 def write_elem(
