@@ -10,7 +10,10 @@ import numpy as np
 import pandas as pd
 from scipy.sparse import issparse
 
+from anndata._io.utils import no_write_dataset_2d
+
 from .._warnings import WriteWarning
+from ..compat import old_positionals
 from ..logging import get_logger
 
 if TYPE_CHECKING:
@@ -21,8 +24,14 @@ if TYPE_CHECKING:
 logger = get_logger(__name__)
 
 
+@no_write_dataset_2d
+@old_positionals("skip_data", "sep")
 def write_csvs(
-    dirname: PathLike, adata: AnnData, skip_data: bool = True, sep: str = ","
+    dirname: PathLike[str] | str,
+    adata: AnnData,
+    *,
+    skip_data: bool = True,
+    sep: str = ",",
 ):
     """See :meth:`~anndata.AnnData.write_csvs`."""
     dirname = Path(dirname)
@@ -47,7 +56,8 @@ def write_csvs(
     for key, value in d_write.items():
         if issparse(value):
             if not_yet_raised_sparse_warning:
-                warnings.warn("Omitting to write sparse annotation.", WriteWarning)
+                msg = "Omitting to write sparse annotation."
+                warnings.warn(msg, WriteWarning, stacklevel=2)
                 not_yet_raised_sparse_warning = False
             continue
         filename = dirname
@@ -61,11 +71,9 @@ def write_csvs(
                 value = value[None]
             try:
                 df = pd.DataFrame(value)
-            except Exception as e:
-                warnings.warn(
-                    f"Omitting to write {key!r} of type {type(e)}.",
-                    WriteWarning,
-                )
+            except Exception as e:  # noqa: BLE001
+                msg = f"Omitting to write {key!r} of type {type(e)}."
+                warnings.warn(msg, WriteWarning, stacklevel=2)
                 continue
         df.to_csv(
             filename,
@@ -75,7 +83,12 @@ def write_csvs(
         )
 
 
-def write_loom(filename: PathLike, adata: AnnData, write_obsm_varm: bool = False):
+@no_write_dataset_2d
+@old_positionals("write_obsm_varm")
+def write_loom(
+    filename: PathLike[str] | str, adata: AnnData, *, write_obsm_varm: bool = False
+) -> None:
+    """See :meth:`~anndata.AnnData.write_loom`."""
     filename = Path(filename)
     row_attrs = {k: np.array(v) for k, v in adata.var.to_dict("list").items()}
     row_names = adata.var_names
@@ -87,13 +100,12 @@ def write_loom(filename: PathLike, adata: AnnData, write_obsm_varm: bool = False
     col_attrs[col_dim] = col_names.values
 
     if adata.X is None:
-        raise ValueError("loompy does not accept empty matrices as data")
+        msg = "loompy does not accept empty matrices as data"
+        raise ValueError(msg)
 
     if write_obsm_varm:
-        for key in adata.obsm.keys():
-            col_attrs[key] = adata.obsm[key]
-        for key in adata.varm.keys():
-            row_attrs[key] = adata.varm[key]
+        col_attrs.update(adata.obsm)
+        row_attrs.update(adata.varm)
     elif len(adata.obsm.keys()) > 0 or len(adata.varm.keys()) > 0:
         logger.warning(
             f"The loom file will lack these fields:\n"
@@ -102,8 +114,8 @@ def write_loom(filename: PathLike, adata: AnnData, write_obsm_varm: bool = False
         )
 
     layers = {"": adata.X.T}
-    for key in adata.layers.keys():
-        layers[key] = adata.layers[key].T
+    for key, layer in adata.layers.items():
+        layers[key] = layer.T
 
     from loompy import create
 
@@ -120,8 +132,8 @@ def _get_chunk_indices(za):
     """
     return [
         (i, j)
-        for i in range(int(math.ceil(float(za.shape[0]) / za.chunks[0])))
-        for j in range(int(math.ceil(float(za.shape[1]) / za.chunks[1])))
+        for i in range(math.ceil(float(za.shape[0]) / za.chunks[0]))
+        for j in range(math.ceil(float(za.shape[1]) / za.chunks[1]))
     ]
 
 
