@@ -9,28 +9,48 @@ import pandas as pd
 import zarr
 from scipy import sparse
 
-from anndata._warnings import OldFormatWarning
-
 from .._core.anndata import AnnData
-from ..compat import _clean_uns, _from_fixed_length_strings
+from .._settings import settings
+from .._warnings import OldFormatWarning
+from ..compat import _clean_uns, _from_fixed_length_strings, is_zarr_v2
 from ..experimental import read_dispatched, write_dispatched
 from .specs import read_elem
-from .utils import _read_legacy_raw, report_read_key_on_error
+from .utils import _read_legacy_raw, no_write_dataset_2d, report_read_key_on_error
 
 if TYPE_CHECKING:
     from collections.abc import MutableMapping
+    from os import PathLike
+
+    from zarr.core.common import AccessModeLiteral
+    from zarr.storage import StoreLike
 
 T = TypeVar("T")
 
 
+def _check_rec_array(adata: AnnData) -> None:
+    if settings.zarr_write_format == 3 and (
+        structured_dtype_keys := {
+            k
+            for k, v in adata.uns.items()
+            if isinstance(v, np.recarray)
+            or (isinstance(v, np.ndarray) and v.dtype.fields)
+        }
+    ):
+        msg = f"zarr v3 does not support structured dtypes.  Found keys {structured_dtype_keys}"
+        raise NotImplementedError(msg)
+
+
+@no_write_dataset_2d
 def write_zarr(
-    store: MutableMapping | str | Path,
+    store: StoreLike,
     adata: AnnData,
     *,
     chunks: tuple[int, ...] | None = None,
     convert_strings_to_categoricals: bool = True,
     **ds_kwargs,
 ) -> None:
+    """See :meth:`~anndata.AnnData.write_zarr`."""
+    _check_rec_array(adata)
     if isinstance(store, Path):
         store = str(store)
     if convert_strings_to_categoricals:
@@ -38,19 +58,29 @@ def write_zarr(
         if adata.raw is not None:
             adata.strings_to_categoricals(adata.raw.var)
     # TODO: Use spec writing system for this
-    f = zarr.open(store, mode="w")
+    f = open_write_group(store)
     f.attrs.setdefault("encoding-type", "anndata")
     f.attrs.setdefault("encoding-version", "0.1.0")
 
-    def callback(func, s, k, elem, dataset_kwargs, iospec):
-        if chunks is not None and not isinstance(elem, sparse.spmatrix) and k == "/X":
+    def callback(
+        write_func, store, elem_name: str, elem, *, dataset_kwargs, iospec
+    ) -> None:
+        if (
+            chunks is not None
+            and not isinstance(elem, sparse.spmatrix)
+            and elem_name.lstrip("/") == "X"
+        ):
             dataset_kwargs = dict(dataset_kwargs, chunks=chunks)
-        func(s, k, elem, dataset_kwargs=dataset_kwargs)
+        write_func(store, elem_name, elem, dataset_kwargs=dataset_kwargs)
 
     write_dispatched(f, "/", adata, callback=callback, dataset_kwargs=ds_kwargs)
+    if is_zarr_v2():
+        zarr.convenience.consolidate_metadata(f.store)
+    else:
+        zarr.consolidate_metadata(f.store)
 
 
-def read_zarr(store: str | Path | MutableMapping | zarr.Group) -> AnnData:
+def read_zarr(store: PathLike[str] | str | MutableMapping | zarr.Group) -> AnnData:
     """\
     Read from a hierarchical Zarr array store.
 
@@ -62,10 +92,7 @@ def read_zarr(store: str | Path | MutableMapping | zarr.Group) -> AnnData:
     if isinstance(store, Path):
         store = str(store)
 
-    if isinstance(store, zarr.Group):
-        f = store
-    else:
-        f = zarr.open(store, mode="r")
+    f = store if isinstance(store, zarr.Group) else zarr.open(store, mode="r")
 
     # Read with handling for backwards compat
     def callback(func, elem_name: str, elem, iospec):
@@ -73,7 +100,7 @@ def read_zarr(store: str | Path | MutableMapping | zarr.Group) -> AnnData:
             return AnnData(
                 **{
                     k: read_dispatched(v, callback)
-                    for k, v in elem.items()
+                    for k, v in dict(elem).items()
                     if not k.startswith("raw.")
                 }
             )
@@ -114,7 +141,7 @@ def read_dataset(dataset: zarr.Array):
     elif len(value.dtype.descr) > 1:  # Compound dtype
         # For backwards compat, now strings are written as variable length
         value = _from_fixed_length_strings(value)
-    if value.shape == ():
+    if value.shape == () and not np.isscalar(value):
         value = value[()]
     return value
 
@@ -123,11 +150,11 @@ def read_dataset(dataset: zarr.Array):
 def read_dataframe_legacy(dataset: zarr.Array) -> pd.DataFrame:
     """Reads old format of dataframes"""
     # NOTE: Likely that categoricals need to be removed from uns
-    warn(
+    msg = (
         f"'{dataset.name}' was written with a very old version of AnnData. "
-        "Consider rewriting it.",
-        OldFormatWarning,
+        "Consider rewriting it."
     )
+    warn(msg, OldFormatWarning, stacklevel=3)
     df = pd.DataFrame(_from_fixed_length_strings(dataset[()]))
     df.set_index(df.columns[0], inplace=True)
     return df
@@ -140,3 +167,11 @@ def read_dataframe(group: zarr.Group | zarr.Array) -> pd.DataFrame:
         return read_dataframe_legacy(group)
     else:
         return read_elem(group)
+
+
+def open_write_group(
+    store: StoreLike, *, mode: AccessModeLiteral = "w", **kwargs
+) -> zarr.Group:
+    if not is_zarr_v2() and "zarr_format" not in kwargs:
+        kwargs["zarr_format"] = settings.zarr_write_format
+    return zarr.open_group(store, mode=mode, **kwargs)
