@@ -3,17 +3,25 @@ from __future__ import annotations
 from types import MappingProxyType
 from typing import TYPE_CHECKING
 
-from .._io.specs import IOSpec, read_elem_as_dask
-from .._types import Read, ReadCallback, StorageType, Write, WriteCallback
+from .._io.specs import IOSpec, read_elem_lazy
+from .._types import (
+    Dataset2DIlocIndexer,
+    Read,
+    ReadCallback,
+    StorageType,
+    Write,
+    WriteCallback,
+)
 from ..utils import module_get_attr_redirect
 from ._dispatch_io import read_dispatched, write_dispatched
+from .backed import read_lazy
 from .merge import concat_on_disk
 from .multi_files import AnnCollection
-from .pytorch import AnnLoader
 
 if TYPE_CHECKING:
     from typing import Any
 
+    from .pytorch import AnnLoader
 
 # Map old name in `anndata.experimental` to new name in `anndata`
 _DEPRECATED = MappingProxyType(
@@ -33,6 +41,10 @@ _DEPRECATED = MappingProxyType(
 
 
 def __getattr__(attr_name: str) -> Any:
+    if attr_name == "AnnLoader":
+        from .pytorch import AnnLoader
+
+        return AnnLoader
     return module_get_attr_redirect(
         attr_name, deprecated_mapping=_DEPRECATED, old_module_path="experimental"
     )
@@ -41,14 +53,16 @@ def __getattr__(attr_name: str) -> Any:
 __all__ = [
     "AnnCollection",
     "AnnLoader",
-    "read_elem_as_dask",
-    "read_dispatched",
-    "write_dispatched",
+    "Dataset2DIlocIndexer",
     "IOSpec",
-    "concat_on_disk",
     "Read",
-    "Write",
     "ReadCallback",
-    "WriteCallback",
     "StorageType",
+    "Write",
+    "WriteCallback",
+    "concat_on_disk",
+    "read_dispatched",
+    "read_elem_lazy",
+    "read_lazy",
+    "write_dispatched",
 ]
