@@ -13,7 +13,7 @@ from scipy import sparse
 import anndata
 
 from ._core.sparse_dataset import BaseCompressedSparseDataset
-from .compat import CupyArray, CupySparseMatrix, DaskArray, SpArray
+from .compat import CSArray, CupyArray, CupySparseMatrix, DaskArray
 from .logging import get_logger
 
 if TYPE_CHECKING:
@@ -23,21 +23,25 @@ if TYPE_CHECKING:
 logger = get_logger(__name__)
 
 
-def import_name(name: str) -> Any:
+def import_name(full_name: str) -> Any:
     from importlib import import_module
 
-    parts = name.split(".")
+    parts = full_name.split(".")
     obj = import_module(parts[0])
-    for i, name in enumerate(parts[1:]):
+    for _i, name in enumerate(parts[1:]):
+        i = _i
         try:
             obj = import_module(f"{obj.__name__}.{name}")
         except ModuleNotFoundError:
             break
+    else:
+        i = len(parts)
     for name in parts[i + 1 :]:
         try:
             obj = getattr(obj, name)
-        except AttributeError:
-            raise RuntimeError(f"{parts[:i]}, {parts[i+1:]}, {obj} {name}")
+        except AttributeError as e:
+            msg = f"{parts[:i]}, {parts[i + 1 :]}, {obj} {name}"
+            raise RuntimeError(msg) from e
     return obj
 
 
@@ -47,7 +51,7 @@ def asarray(x):
     return np.asarray(x)
 
 
-@asarray.register(SpArray)
+@asarray.register(CSArray)
 @asarray.register(sparse.spmatrix)
 def asarray_sparse(x):
     return x.toarray()
@@ -91,11 +95,12 @@ def convert_to_dict_dict(obj: dict):
 @convert_to_dict.register(np.ndarray)
 def convert_to_dict_ndarray(obj: np.ndarray):
     if obj.dtype.fields is None:
-        raise TypeError(
+        msg = (
             "Can only convert np.ndarray with compound dtypes to dict, "
             f"passed array had “{obj.dtype}”."
         )
-    return {k: obj[k] for k in obj.dtype.fields.keys()}
+        raise TypeError(msg)
+    return {k: obj[k] for k in obj.dtype.fields}
 
 
 @convert_to_dict.register(type(None))
@@ -121,13 +126,11 @@ try:
         if layout.is_numpy:
             # if it's an embedded rectilinear array, we have to deal with its shape
             # which might not be 1-dimensional
-            if layout.is_unknown:
-                shape = (0,)
-            else:
-                shape = layout.shape
+            shape = (0,) if layout.is_unknown else layout.shape
             numpy_axis = lateral_context["axis"] - depth + 1
             if not (1 <= numpy_axis < len(shape)):
-                raise TypeError(f"axis={lateral_context['axis']} is too deep")
+                msg = f"axis={lateral_context['axis']} is too deep"
+                raise TypeError(msg)
             lateral_context["out"] = shape[numpy_axis]
             return ak.contents.EmptyArray()
 
@@ -136,7 +139,8 @@ try:
                 # Strings are implemented like an array of lists of uint8 (ListType(NumpyType(...)))
                 # which results in an extra hierarchy-level that shouldn't show up in dim_len
                 # See https://github.com/scikit-hep/awkward/discussions/1654#discussioncomment-3736747
-                raise TypeError(f"axis={lateral_context['axis']} is too deep")
+                msg = f"axis={lateral_context['axis']} is too deep"
+                raise TypeError(msg)
 
             if layout.is_regular:
                 # if it's a regular list, you want the size
@@ -154,9 +158,8 @@ try:
             # currently, we don't recurse into records
             # in theory we could, just not sure how to do it at the moment
             # Would need to consider cases like: scalars, unevenly sized values
-            raise TypeError(
-                f"Cannot recurse into record type found at axis={lateral_context['axis']}"
-            )
+            msg = f"Cannot recurse into record type found at axis={lateral_context['axis']}"
+            raise TypeError(msg)
 
         elif layout.is_union:
             # if it's a union, you could get the result of each union branch
@@ -187,7 +190,8 @@ try:
         Code adapted from @jpivarski's solution in https://github.com/scikit-hep/awkward/discussions/1654#discussioncomment-3521574
         """
         if axis < 0:  # negative axis is another can of worms... maybe later
-            raise NotImplementedError("Does not support negative axis")
+            msg = "Does not support negative axis"
+            raise NotImplementedError(msg)
         elif axis == 0:
             return len(array)
         else:
@@ -261,14 +265,15 @@ def make_index_unique(index: pd.Index, join: str = "-"):
                 example_colliding_values.append(tentative_new_name)
 
     if issue_interpretation_warning:
-        warnings.warn(
-            f"Suffix used ({join}[0-9]+) to deduplicate index values may make index "
-            + "values difficult to interpret. There values with a similar suffixes in "
-            + "the index. Consider using a different delimiter by passing "
-            + "`join={delimiter}`"
-            + "Example key collisions generated by the make_index_unique algorithm: "
-            + str(example_colliding_values)
+        msg = (
+            f"Suffix used ({join}[0-9]+) to deduplicate index values may make index values difficult to interpret. "
+            "There values with a similar suffixes in the index. "
+            "Consider using a different delimiter by passing `join={delimiter}`. "
+            "Example key collisions generated by the make_index_unique algorithm: "
+            f"{example_colliding_values}"
         )
+        # 3: caller -> 2: `{obs,var}_names_make_unique` -> 1: here
+        warnings.warn(msg, UserWarning, stacklevel=3)
     values[indices_dup] = values_dup
     index = pd.Index(values, name=index.name)
     return index
@@ -304,7 +309,9 @@ def ensure_df_homogeneous(
     else:
         arr = df.to_numpy()
     if df.dtypes.nunique() != 1:
-        warnings.warn(f"{name} converted to numpy array with dtype {arr.dtype}")
+        msg = f"{name} converted to numpy array with dtype {arr.dtype}"
+        # 4: caller -> 3: `AnnData.__init__` -> 2: `_init_as_actual` → 1: here
+        warnings.warn(msg, UserWarning, stacklevel=4)
     return arr
 
 
@@ -317,11 +324,12 @@ def convert_dictionary_to_structured_array(source: Mapping[str, Sequence[Any]]):
             else np.asarray(col).astype("U")
             for col in source.values()
         ]
-    except UnicodeEncodeError:
-        raise ValueError(
+    except UnicodeEncodeError as e:
+        msg = (
             "Currently only support ascii strings. "
             "Don’t use “ö” etc. for sample annotation."
         )
+        raise ValueError(msg) from e
 
     # if old_index_key not in source:
     #     names.append(new_index_key)
@@ -330,7 +338,12 @@ def convert_dictionary_to_structured_array(source: Mapping[str, Sequence[Any]]):
     #     names[names.index(old_index_key)] = new_index_key
     #     cols[names.index(old_index_key)] = cols[names.index(old_index_key)].astype("U")
     dtype_list = list(
-        zip(names, [str(c.dtype) for c in cols], [(c.shape[1],) for c in cols])
+        zip(
+            names,
+            [str(c.dtype) for c in cols],
+            [(c.shape[1],) for c in cols],
+            strict=True,
+        )
     )
     # might be unnecessary
     dtype = np.dtype(dtype_list)
@@ -354,13 +367,14 @@ def warn_once(msg: str, category: type[Warning], stacklevel: int = 1):
 
 def deprecated(
     new_name: str,
-    category: type[Warning] = DeprecationWarning,
+    category: type[Warning] = FutureWarning,
     add_msg: str = "",
+    *,
     hide: bool = True,
 ):
     """\
     This is a decorator which can be used to mark functions
-    as deprecated. It will result in a warning being emitted
+    as deprecated with a FutureWarning. It will result in a warning being emitted
     when the function is used.
     """
 
@@ -426,7 +440,7 @@ def module_get_attr_redirect(
             f"Importing {attr_name} from `{full_old_module_path}` is deprecated. "
             f"Import anndata.{new_path} instead."
         )
-        warnings.warn(msg, FutureWarning)
+        warnings.warn(msg, FutureWarning, stacklevel=2)
         # hacky import_object_by_name, but we test all these
         mod = anndata
         while "." in new_path:
