@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import warnings
 from collections.abc import Hashable
-from contextlib import nullcontext
 from copy import deepcopy
 from functools import partial, singledispatch
+from importlib.util import find_spec
 from itertools import chain, permutations, product
 from operator import attrgetter
 from typing import TYPE_CHECKING
@@ -12,6 +12,7 @@ from typing import TYPE_CHECKING
 import numpy as np
 import pandas as pd
 import pytest
+import scipy
 from boltons.iterutils import default_exit, remap, research
 from numpy import ma
 from packaging.version import Version
@@ -20,7 +21,8 @@ from scipy import sparse
 from anndata import AnnData, Raw, concat
 from anndata._core import merge
 from anndata._core.index import _subset
-from anndata.compat import AwkArray, CupySparseMatrix, DaskArray, SpArray
+from anndata._core.xarray import Dataset2D
+from anndata.compat import AwkArray, CSArray, CSMatrix, CupySparseMatrix, DaskArray
 from anndata.tests import helpers
 from anndata.tests.helpers import (
     BASE_MATRIX_PARAMS,
@@ -61,7 +63,7 @@ def _filled_array(a, fill_value=None):
     return as_dense_dask_array(_filled_array_np(a, fill_value))
 
 
-@filled_like.register(sparse.spmatrix)
+@filled_like.register(CSMatrix)
 def _filled_sparse(a, fill_value=None):
     if fill_value is None:
         return sparse.csr_matrix(a.shape)
@@ -69,7 +71,7 @@ def _filled_sparse(a, fill_value=None):
         return sparse.csr_matrix(np.broadcast_to(fill_value, a.shape))
 
 
-@filled_like.register(SpArray)
+@filled_like.register(CSArray)
 def _filled_sparse_array(a, fill_value=None):
     return sparse.csr_array(filled_like(sparse.csr_matrix(a)))
 
@@ -125,7 +127,35 @@ def merge_strategy(request):
     return request.param
 
 
-def fix_known_differences(orig, result, backwards_compat=True):
+@pytest.fixture(
+    params=[
+        pytest.param(False, id="pandas"),
+        pytest.param(
+            True,
+            marks=pytest.mark.skipif(
+                not find_spec("xarray"), reason="xarray not installed."
+            ),
+            id="xarray",
+        ),
+    ],
+)
+def use_xdataset(request):
+    return request.param
+
+
+@pytest.fixture(
+    params=[
+        pytest.param(False, id="concat-in-memory"),
+        pytest.param(True, id="concat-lazy"),
+    ],
+)
+def force_lazy(request):
+    return request.param
+
+
+def fix_known_differences(
+    orig: AnnData, result: AnnData, *, backwards_compat: bool = True
+):
     """
     Helper function for reducing anndata's to only the elements we expect to be
     equivalent after concatenation.
@@ -137,6 +167,34 @@ def fix_known_differences(orig, result, backwards_compat=True):
     orig = orig.copy()
     result = result.copy()
 
+    if backwards_compat:
+        del orig.varm
+        del orig.varp
+        if isinstance(result.obs, Dataset2D):
+            result.obs = result.obs.ds.drop_vars(["batch"])
+        else:
+            result.obs.drop(columns=["batch"], inplace=True)
+
+    for attrname in ("obs", "var"):
+        if isinstance(getattr(result, attrname), Dataset2D):
+            for adata in (orig, result):
+                df = getattr(adata, attrname).ds.to_dataframe()[
+                    getattr(orig, attrname).columns
+                ]
+                df.index.name = "index"
+                setattr(adata, attrname, df)
+            resattr = getattr(result, attrname)
+            origattr = getattr(orig, attrname)
+            for colname, col in resattr.items():
+                # concatenation of XDatasets happens via Dask arrays and those don't know about Pandas Extension arrays
+                # so categoricals and nullable arrays are all converted to other dtypes
+                if col.dtype != origattr[
+                    colname
+                ].dtype and pd.api.types.is_extension_array_dtype(
+                    origattr[colname].dtype
+                ):
+                    resattr[colname] = col.astype(origattr[colname].dtype)
+
     result.strings_to_categoricals()  # Should this be implicit in concatenation?
 
     # TODO
@@ -144,21 +202,21 @@ def fix_known_differences(orig, result, backwards_compat=True):
     # * merge obsp, but some information should be lost
     del orig.obsp  # TODO
 
-    if backwards_compat:
-        del orig.varm
-        del orig.varp
-        result.obs.drop(columns=["batch"], inplace=True)
-
     # Possibly need to fix this, ordered categoricals lose orderedness
-    for k, dtype in orig.obs.dtypes.items():
-        if isinstance(dtype, pd.CategoricalDtype) and dtype.ordered:
-            result.obs[k] = result.obs[k].astype(dtype)
+    for get_df in [lambda k: k.obs, lambda k: k.obsm["df"]]:
+        str_to_df_converted = get_df(result)
+        for k, dtype in get_df(orig).dtypes.items():
+            if isinstance(dtype, pd.CategoricalDtype) and dtype.ordered:
+                str_to_df_converted[k] = str_to_df_converted[k].astype(dtype)
 
     return orig, result
 
 
-def test_concat_interface_errors():
-    adatas = [gen_adata((5, 10)), gen_adata((5, 10))]
+def test_concat_interface_errors(use_xdataset):
+    adatas = [
+        gen_adata((5, 10), obs_xdataset=use_xdataset, var_xdataset=use_xdataset),
+        gen_adata((5, 10), obs_xdataset=use_xdataset, var_xdataset=use_xdataset),
+    ]
 
     with pytest.raises(ValueError, match="`axis` must be.*0, 1, 'obs', or 'var'"):
         concat(adatas, axis=3)
@@ -176,8 +234,23 @@ def test_concat_interface_errors():
         (lambda x, **kwargs: x[0].concatenate(x[1:], **kwargs), True),
     ],
 )
-def test_concatenate_roundtrip(join_type, array_type, concat_func, backwards_compat):
-    adata = gen_adata((100, 10), X_type=array_type, **GEN_ADATA_DASK_ARGS)
+def test_concatenate_roundtrip(
+    join_type,
+    array_type,
+    concat_func,
+    backwards_compat,
+    use_xdataset,
+    force_lazy,
+):
+    if backwards_compat and force_lazy:
+        pytest.skip("unsupported")
+    adata = gen_adata(
+        (100, 10),
+        X_type=array_type,
+        obs_xdataset=use_xdataset,
+        var_xdataset=use_xdataset,
+        **GEN_ADATA_DASK_ARGS,
+    )
 
     remaining = adata.obs_names
     subsets = []
@@ -186,8 +259,13 @@ def test_concatenate_roundtrip(join_type, array_type, concat_func, backwards_com
         subset_idx = np.random.choice(remaining, n, replace=False)
         subsets.append(adata[subset_idx])
         remaining = remaining.difference(subset_idx)
-
     result = concat_func(subsets, join=join_type, uns_merge="same", index_unique=None)
+    if backwards_compat and use_xdataset:
+        import xarray as xr
+
+        result.var = xr.Dataset.from_dataframe(
+            result.var
+        )  # backwards compat always returns a dataframe
 
     # Correcting for known differences
     orig, result = fix_known_differences(
@@ -197,10 +275,7 @@ def test_concatenate_roundtrip(join_type, array_type, concat_func, backwards_com
     assert_equal(result[orig.obs_names].copy(), orig)
     base_type = type(orig.X)
     if sparse.issparse(orig.X):
-        if isinstance(orig.X, SpArray):
-            base_type = SpArray
-        else:
-            base_type = sparse.spmatrix
+        base_type = CSArray if isinstance(orig.X, CSArray) else CSMatrix
     if isinstance(orig.X, CupySparseMatrix):
         base_type = CupySparseMatrix
     assert isinstance(result.X, base_type)
@@ -404,7 +479,7 @@ def test_concatenate_obsm_outer(obsm_adatas, fill_val):
         ),
     )
 
-    assert isinstance(outer.obsm["sparse"], sparse.spmatrix)
+    assert isinstance(outer.obsm["sparse"], CSMatrix)
     np.testing.assert_equal(
         outer.obsm["sparse"].toarray(),
         np.array(
@@ -495,19 +570,19 @@ def test_concatenate_fill_value(fill_val):
     adata1.obsm = {
         k: v
         for k, v in adata1.obsm.items()
-        if not isinstance(v, pd.DataFrame | AwkArray)
+        if not isinstance(v, pd.DataFrame | AwkArray | Dataset2D)
     }
     adata2 = gen_adata((10, 5))
     adata2.obsm = {
         k: v[:, : v.shape[1] // 2]
         for k, v in adata2.obsm.items()
-        if not isinstance(v, pd.DataFrame | AwkArray)
+        if not isinstance(v, pd.DataFrame | AwkArray | Dataset2D)
     }
     adata3 = gen_adata((7, 3))
     adata3.obsm = {
         k: v[:, : v.shape[1] // 3]
         for k, v in adata3.obsm.items()
-        if not isinstance(v, pd.DataFrame | AwkArray)
+        if not isinstance(v, pd.DataFrame | AwkArray | Dataset2D)
     }
     # remove AwkArrays from adata.var, as outer joins are not yet implemented for them
     for tmp_ad in [adata1, adata2, adata3]:
@@ -853,7 +928,7 @@ def test_pairwise_concat(axis_name, array_type):
             obsp={"arr": gen_axis_array(m)},
             varp={"arr": gen_axis_array(n)},
         )
-        for k, m, n in zip("abc", Ms, Ns)
+        for k, m, n in zip("abc", Ms, Ns, strict=True)
     }
 
     w_pairwise = concat(adatas, axis=axis, label="orig", pairwise=True)
@@ -1044,7 +1119,9 @@ def gen_list(n):
 
 
 def gen_sparse(n):
-    return sparse.random(np.random.randint(1, 100), np.random.randint(1, 100))
+    return sparse.random(
+        np.random.randint(1, 100), np.random.randint(1, 100), format="csr"
+    )
 
 
 def gen_something(n):
@@ -1150,35 +1227,76 @@ def test_concatenate_uns(unss, merge_strategy, result, value_gen):
     """
     # So we can see what the initial pattern was meant to be
     print(merge_strategy, "\n", unss, "\n", result)
-    result, *unss = permute_nested_values([result] + unss, value_gen)
+    result, *unss = permute_nested_values([result, *unss], value_gen)
     adatas = [uns_ad(uns) for uns in unss]
     with pytest.warns(FutureWarning, match=r"concatenate is deprecated"):
         merged = AnnData.concatenate(*adatas, uns_merge=merge_strategy).uns
     assert_equal(merged, result, elem_name="uns")
 
 
-def test_transposed_concat(array_type, axis_name, join_type, merge_strategy):
+def test_transposed_concat(
+    array_type,
+    axis_name,
+    join_type,
+    merge_strategy,
+    use_xdataset,
+    force_lazy,
+):
     axis, axis_name = merge._resolve_axis(axis_name)
     alt_axis = 1 - axis
-    lhs = gen_adata((10, 10), X_type=array_type, **GEN_ADATA_DASK_ARGS)
-    rhs = gen_adata((10, 12), X_type=array_type, **GEN_ADATA_DASK_ARGS)
+    lhs = gen_adata(
+        (10, 10),
+        X_type=array_type,
+        obs_xdataset=use_xdataset,
+        var_xdataset=use_xdataset,
+        **GEN_ADATA_DASK_ARGS,
+    )
+    rhs = gen_adata(
+        (10, 12),
+        X_type=array_type,
+        obs_xdataset=use_xdataset,
+        var_xdataset=use_xdataset,
+        **GEN_ADATA_DASK_ARGS,
+    )
 
-    a = concat([lhs, rhs], axis=axis, join=join_type, merge=merge_strategy)
-    b = concat([lhs.T, rhs.T], axis=alt_axis, join=join_type, merge=merge_strategy).T
+    a = concat(
+        [lhs, rhs],
+        axis=axis,
+        join=join_type,
+        merge=merge_strategy,
+        force_lazy=force_lazy,
+    )
+    b = concat(
+        [lhs.T, rhs.T],
+        axis=alt_axis,
+        join=join_type,
+        merge=merge_strategy,
+        force_lazy=force_lazy,
+    ).T
 
     assert_equal(a, b)
 
 
-def test_batch_key(axis_name):
+def test_batch_key(axis_name, use_xdataset, force_lazy):
     """Test that concat only adds a label if the key is provided"""
 
     get_annot = attrgetter(axis_name)
 
-    lhs = gen_adata((10, 10), **GEN_ADATA_DASK_ARGS)
-    rhs = gen_adata((10, 12), **GEN_ADATA_DASK_ARGS)
+    lhs = gen_adata(
+        (10, 10),
+        obs_xdataset=use_xdataset,
+        var_xdataset=use_xdataset,
+        **GEN_ADATA_DASK_ARGS,
+    )
+    rhs = gen_adata(
+        (10, 12),
+        obs_xdataset=use_xdataset,
+        var_xdataset=use_xdataset,
+        **GEN_ADATA_DASK_ARGS,
+    )
 
     # There is probably a prettier way to do this
-    annot = get_annot(concat([lhs, rhs], axis=axis_name))
+    annot = get_annot(concat([lhs, rhs], axis=axis_name, force_lazy=force_lazy))
     assert (
         list(
             annot.columns.difference(
@@ -1188,7 +1306,9 @@ def test_batch_key(axis_name):
         == []
     )
 
-    batch_annot = get_annot(concat([lhs, rhs], axis=axis_name, label="batch"))
+    batch_annot = get_annot(
+        concat([lhs, rhs], axis=axis_name, label="batch", force_lazy=force_lazy)
+    )
     assert list(
         batch_annot.columns.difference(
             get_annot(lhs).columns.union(get_annot(rhs).columns)
@@ -1196,16 +1316,16 @@ def test_batch_key(axis_name):
     ) == ["batch"]
 
 
-def test_concat_categories_from_mapping():
+def test_concat_categories_from_mapping(use_xdataset, force_lazy):
     mapping = {
-        "a": gen_adata((10, 10)),
-        "b": gen_adata((10, 10)),
+        "a": gen_adata((10, 10), obs_xdataset=use_xdataset, var_xdataset=use_xdataset),
+        "b": gen_adata((10, 10), obs_xdataset=use_xdataset, var_xdataset=use_xdataset),
     }
     keys = list(mapping.keys())
     adatas = list(mapping.values())
 
-    mapping_call = partial(concat, mapping)
-    iter_call = partial(concat, adatas, keys=keys)
+    mapping_call = partial(concat, mapping, force_lazy=force_lazy)
+    iter_call = partial(concat, adatas, keys=keys, force_lazy=force_lazy)
 
     assert_equal(mapping_call(), iter_call())
     assert_equal(mapping_call(label="batch"), iter_call(label="batch"))
@@ -1249,9 +1369,9 @@ def test_concat_categories_maintain_dtype():
 
     result = concat({"a": a, "b": b, "c": c}, join="outer")
 
-    assert isinstance(
-        result.obs["cat"].dtype, pd.CategoricalDtype
-    ), f"Was {result.obs['cat'].dtype}"
+    assert isinstance(result.obs["cat"].dtype, pd.CategoricalDtype), (
+        f"Was {result.obs['cat'].dtype}"
+    )
     assert pd.api.types.is_string_dtype(result.obs["cat_ordered"])
 
 
@@ -1335,15 +1455,17 @@ def test_bool_promotion():
     assert result.obs["bool"].dtype == np.dtype(bool)
 
 
-def test_concat_names(axis_name):
+def test_concat_names(axis_name, use_xdataset, force_lazy):
     get_annot = attrgetter(axis_name)
 
-    lhs = gen_adata((10, 10))
-    rhs = gen_adata((10, 10))
+    lhs = gen_adata((10, 10), obs_xdataset=use_xdataset, var_xdataset=use_xdataset)
+    rhs = gen_adata((10, 10), obs_xdataset=use_xdataset, var_xdataset=use_xdataset)
 
-    assert not get_annot(concat([lhs, rhs], axis=axis_name)).index.is_unique
+    assert not get_annot(
+        concat([lhs, rhs], axis=axis_name, force_lazy=force_lazy)
+    ).index.is_unique
     assert get_annot(
-        concat([lhs, rhs], axis=axis_name, index_unique="-")
+        concat([lhs, rhs], axis=axis_name, index_unique="-", force_lazy=force_lazy)
     ).index.is_unique
 
 
@@ -1372,33 +1494,39 @@ def expected_shape(
 @pytest.mark.parametrize(
     "shape", [pytest.param((8, 0), id="no_var"), pytest.param((0, 10), id="no_obs")]
 )
-def test_concat_size_0_axis(axis_name, join_type, merge_strategy, shape):
+def test_concat_size_0_axis(
+    axis_name, join_type, merge_strategy, shape, use_xdataset, force_lazy
+):
     """Regression test for https://github.com/scverse/anndata/issues/526"""
     axis, axis_name = merge._resolve_axis(axis_name)
     alt_axis = 1 - axis
     col_dtypes = (*DEFAULT_COL_TYPES, pd.StringDtype)
-    a = gen_adata((5, 7), obs_dtypes=col_dtypes, var_dtypes=col_dtypes)
-    b = gen_adata(shape, obs_dtypes=col_dtypes, var_dtypes=col_dtypes)
+    a = gen_adata(
+        (5, 7),
+        obs_dtypes=col_dtypes,
+        var_dtypes=col_dtypes,
+        obs_xdataset=use_xdataset,
+        var_xdataset=use_xdataset,
+    )
+    b = gen_adata(
+        shape,
+        obs_dtypes=col_dtypes,
+        var_dtypes=col_dtypes,
+        obs_xdataset=use_xdataset,
+        var_xdataset=use_xdataset,
+    )
 
     expected_size = expected_shape(a, b, axis=axis, join=join_type)
 
-    ctx_concat_empty = (
-        pytest.warns(
-            FutureWarning,
-            match=r"The behavior of DataFrame concatenation with empty or all-NA entries is deprecated",
-        )
-        if shape[axis] == 0 and Version(pd.__version__) >= Version("2.1")
-        else nullcontext()
+    result = concat(
+        {"a": a, "b": b},
+        axis=axis,
+        join=join_type,
+        merge=merge_strategy,
+        pairwise=True,
+        index_unique="-",
+        force_lazy=force_lazy,
     )
-    with ctx_concat_empty:
-        result = concat(
-            {"a": a, "b": b},
-            axis=axis,
-            join=join_type,
-            merge=merge_strategy,
-            pairwise=True,
-            index_unique="-",
-        )
     assert result.shape == expected_size
 
     if join_type == "outer":
@@ -1410,7 +1538,7 @@ def test_concat_size_0_axis(axis_name, join_type, merge_strategy, shape):
 
         check_filled_like(result.X[axis_idx], elem_name="X")
         check_filled_like(result.X[altaxis_idx], elem_name="X")
-        for k, elem in getattr(result, "layers").items():
+        for k, elem in result.layers.items():
             check_filled_like(elem[axis_idx], elem_name=f"layers/{k}")
             check_filled_like(elem[altaxis_idx], elem_name=f"layers/{k}")
 
@@ -1436,15 +1564,33 @@ def test_concat_size_0_axis(axis_name, join_type, merge_strategy, shape):
 
 
 @pytest.mark.parametrize("elem", ["sparse", "array", "df", "da"])
-def test_concat_outer_aligned_mapping(elem):
-    a = gen_adata((5, 5), **GEN_ADATA_DASK_ARGS)
-    b = gen_adata((3, 5), **GEN_ADATA_DASK_ARGS)
-    del b.obsm[elem]
+@pytest.mark.parametrize("axis", ["obs", "var"])
+def test_concat_outer_aligned_mapping(elem, axis, use_xdataset, force_lazy):
+    a = gen_adata(
+        (5, 5),
+        obs_xdataset=use_xdataset,
+        var_xdataset=use_xdataset,
+        **GEN_ADATA_DASK_ARGS,
+    )
+    b = gen_adata(
+        (3, 5),
+        obs_xdataset=use_xdataset,
+        var_xdataset=use_xdataset,
+        **GEN_ADATA_DASK_ARGS,
+    )
+    del getattr(b, f"{axis}m")[elem]
 
-    concated = concat({"a": a, "b": b}, join="outer", label="group")
-    result = concated.obsm[elem][concated.obs["group"] == "b"]
+    concated = concat(
+        {"a": a, "b": b}, join="outer", label="group", axis=axis, force_lazy=force_lazy
+    )
 
-    check_filled_like(result, elem_name=f"obsm/{elem}")
+    mask = getattr(concated, axis)["group"] == "b"
+    result = getattr(
+        concated[(mask, slice(None)) if axis == "obs" else (slice(None), mask)],
+        f"{axis}m",
+    )[elem]
+
+    check_filled_like(result, elem_name=f"{axis}m/{elem}")
 
 
 @mark_legacy_concatenate
@@ -1455,12 +1601,15 @@ def test_concatenate_size_0_axis():
     b = gen_adata((5, 0))
 
     # Mostly testing that this doesn't error
-    a.concatenate([b]).shape == (10, 0)
-    b.concatenate([a]).shape == (10, 0)
+    assert a.concatenate([b]).shape == (10, 0)
+    assert b.concatenate([a]).shape == (10, 0)
 
 
-def test_concat_null_X():
-    adatas_orig = {k: gen_adata((20, 10)) for k in list("abc")}
+def test_concat_null_X(use_xdataset):
+    adatas_orig = {
+        k: gen_adata((20, 10), obs_xdataset=use_xdataset, var_xdataset=use_xdataset)
+        for k in list("abc")
+    }
     adatas_no_X = {}
     for k, v in adatas_orig.items():
         v = v.copy()
@@ -1492,14 +1641,18 @@ def test_concat_X_dtype(cpu_array_type, sparse_indexer_type):
     assert result.X.dtype == np.int8
     assert result.raw.X.dtype == np.float64
     if sparse.issparse(result.X):
-        # See https://github.com/scipy/scipy/issues/20389 for why this doesn't work with csc
+        # https://github.com/scipy/scipy/issues/20389 was merged in 1.15 but is still an issue with matrix
         if sparse_indexer_type == np.int64 and (
-            issubclass(cpu_array_type, sparse.spmatrix) or adata.X.format == "csc"
+            (
+                (issubclass(cpu_array_type, CSArray) or adata.X.format == "csc")
+                and Version(scipy.__version__) < Version("1.15.0")
+            )
+            or issubclass(cpu_array_type, CSMatrix)
         ):
             pytest.xfail(
                 "Data type int64 is not maintained for sparse matrices or csc array"
             )
-        assert result.X.indptr.dtype == sparse_indexer_type
+        assert result.X.indptr.dtype == sparse_indexer_type, result.X
         assert result.X.indices.dtype == sparse_indexer_type
 
 
@@ -1529,6 +1682,35 @@ def test_concat_different_types_dask(merge_strategy, array_type):
 
     assert_equal(result1, target1)
     assert_equal(result2, target2)
+
+
+def test_concat_missing_elem_dask_join(join_type):
+    import dask.array as da
+
+    import anndata as ad
+
+    ad1 = ad.AnnData(X=np.ones((5, 10)))
+    ad2 = ad.AnnData(X=np.zeros((5, 5)), layers={"a": da.ones((5, 5))})
+    ad_in_memory_with_layers = ad2.to_memory()
+
+    result1 = ad.concat([ad1, ad2], join=join_type)
+    result2 = ad.concat([ad1, ad_in_memory_with_layers], join=join_type)
+    assert_equal(result1, result2)
+
+
+def test_impute_dask(axis_name):
+    import dask.array as da
+
+    from anndata._core.merge import _resolve_axis, missing_element
+
+    axis, _ = _resolve_axis(axis_name)
+    els = [da.ones((5, 5))]
+    missing = missing_element(6, els, axis=axis, off_axis_size=17)
+    assert isinstance(missing, DaskArray)
+    in_memory = missing.compute()
+    assert np.all(np.isnan(in_memory))
+    assert in_memory.shape[axis] == 6
+    assert in_memory.shape[axis - 1] == 17
 
 
 def test_outer_concat_with_missing_value_for_df():
@@ -1643,7 +1825,7 @@ def test_concat_dask_sparse_matches_memory(join_type, merge_strategy):
     X = sparse.random(50, 20, density=0.5, format="csr")
     X_dask = da.from_array(X, chunks=(5, 20))
     var_names_1 = [f"gene_{i}" for i in range(20)]
-    var_names_2 = [f"gene_{i}{'_foo' if (i%2) else ''}" for i in range(20, 40)]
+    var_names_2 = [f"gene_{i}{'_foo' if (i % 2) else ''}" for i in range(20, 40)]
 
     ad1 = AnnData(X=X, var=pd.DataFrame(index=var_names_1))
     ad2 = AnnData(X=X, var=pd.DataFrame(index=var_names_2))
@@ -1655,3 +1837,9 @@ def test_concat_dask_sparse_matches_memory(join_type, merge_strategy):
     res_dask = concat([ad1_dask, ad2_dask], join=join_type, merge=merge_strategy)
 
     assert_equal(res_in_memory, res_dask)
+
+
+def test_1d_concat():
+    adata = AnnData(np.ones((5, 20)), obsm={"1d-array": np.ones(5)})
+    concated = concat([adata, adata])
+    assert concated.obsm["1d-array"].shape == (10, 1)
