@@ -12,10 +12,7 @@ from anndata import AnnData, concat
 from anndata._core.merge import _resolve_axis
 from anndata.experimental.merge import as_group, concat_on_disk
 from anndata.io import read_elem, write_elem
-from anndata.tests.helpers import (
-    assert_equal,
-    gen_adata,
-)
+from anndata.tests.helpers import assert_equal, gen_adata
 from anndata.utils import asarray
 
 if TYPE_CHECKING:
@@ -30,7 +27,7 @@ GEN_ADATA_OOC_CONCAT_ARGS = dict(
         pd.DataFrame,
     ),
     varm_types=(sparse.csr_matrix, np.ndarray, pd.DataFrame),
-    layers_types=(sparse.spmatrix, np.ndarray, pd.DataFrame),
+    layers_types=(sparse.csr_matrix, np.ndarray, pd.DataFrame),
 )
 
 
@@ -54,8 +51,8 @@ def file_format(request) -> Literal["zarr", "h5ad"]:
     return request.param
 
 
-# trying with 10 should be slow but will guarantee that the feature is being used
-@pytest.fixture(params=[10, 100_000_000])
+# 1000 is enough to guarantee that the feature is being used
+@pytest.fixture(params=[1_000, 100_000_000])
 def max_loaded_elems(request) -> int:
     return request.param
 
@@ -97,7 +94,7 @@ def assert_eq_concat_on_disk(
     if max_loaded_elems is not None:
         kwargs["max_loaded_elems"] = max_loaded_elems
     concat_on_disk(paths, out_name, *args, **kwargs)
-    res2 = read_elem(as_group(out_name))
+    res2 = read_elem(as_group(out_name, mode="r"))
     assert_equal(res1, res2, exact=False)
 
 
@@ -108,11 +105,13 @@ def get_array_type(array_type, axis):
         return sparse.csr_array if axis == 0 else sparse.csc_array
     if array_type == "array":
         return asarray
-    raise NotImplementedError(f"array_type {array_type} not implemented")
+    msg = f"array_type {array_type} not implemented"
+    raise NotImplementedError(msg)
 
 
 @pytest.mark.parametrize("reindex", [True, False], ids=["reindex", "no_reindex"])
 def test_anndatas(
+    *,
     axis: Literal[0, 1],
     array_type: Literal["array", "sparse", "sparse_array"],
     join_type: Literal["inner", "outer"],
