@@ -8,12 +8,7 @@ import pytest
 from scipy import sparse
 
 import anndata as ad
-from anndata.compat import (
-    CupyArray,
-    CupyCSRMatrix,
-    DaskArray,
-    add_note,
-)
+from anndata.compat import CupyArray, CupyCSRMatrix, DaskArray
 from anndata.tests.helpers import (
     BASE_MATRIX_PARAMS,
     CUPY_MATRIX_PARAMS,
@@ -107,7 +102,8 @@ def test_gen_random_column(dtype):
 # Does this work for every warning?
 def test_report_name():
     def raise_error():
-        raise Exception("an error occurred!")
+        msg = "an error occurred!"
+        raise Exception(msg)
 
     letters = np.array(list(ascii_letters))
     tag = "".join(np.random.permutation(letters))
@@ -144,7 +140,7 @@ def test_assert_equal():
     #     exact=False,
     # )
     adata2 = adata.copy()
-    to_modify = list(adata2.layers.keys())[0]
+    to_modify = next(iter(adata2.layers.keys()))
     del adata2.layers[to_modify]
     with pytest.raises(AssertionError) as missing_layer_error:
         assert_equal(adata, adata2)
@@ -276,33 +272,6 @@ def test_assert_equal_dask_sparse_arrays():
 
     assert_equal(x, y)
     assert_equal(y, x)
-
-
-@pytest.mark.parametrize(
-    ("error", "match"),
-    [
-        (Exception("test"), "test"),
-        (add_note(AssertionError("foo"), "bar"), "bar"),
-        (add_note(add_note(AssertionError("foo"), "bar"), "baz"), "bar"),
-        (add_note(add_note(AssertionError("foo"), "bar"), "baz"), "baz"),
-    ],
-)
-def test_check_error_notes_success(error, match):
-    with pytest.raises(Exception, match=match):
-        raise error
-
-
-@pytest.mark.parametrize(
-    ("error", "match"),
-    [
-        (Exception("test"), "foo"),
-        (add_note(AssertionError("foo"), "bar"), "baz"),
-    ],
-)
-def test_check_error_notes_failure(error, match):
-    with pytest.raises(AssertionError):
-        with pytest.raises(Exception, match=match):
-            raise error
 
 
 @pytest.mark.parametrize(
