@@ -10,7 +10,7 @@ import pytest
 from packaging.version import Version
 
 import anndata as ad
-from anndata.tests.helpers import gen_adata
+from anndata.tests.helpers import GEN_ADATA_NO_XARRAY_ARGS, gen_adata
 
 
 @pytest.mark.skipif(not find_spec("scanpy"), reason="Scanpy is not installed")
@@ -21,30 +21,24 @@ def test_old_format_warning_thrown():
         )
         import scanpy as sc
 
+    def msg_re(entry: str) -> str:
+        return re.escape(
+            f"Moving element from .uns['neighbors'][{entry!r}] to .obsp[{entry!r}]."
+        )
+
     pth = Path(sc.datasets.__file__).parent / "10x_pbmc68k_reduced.h5ad"
-    # TODO: with Pytest 8, all this can be a
-    #       `with pytest.warns(...), pytest.warns(...):`
-    with warnings.catch_warnings(record=True) as record:
-        warnings.simplefilter("always", ad.OldFormatWarning)
-        warnings.simplefilter("always", FutureWarning)
+    warnings.simplefilter("default", FutureWarning)
+    with (
+        pytest.warns(FutureWarning, match=msg_re("distances")),
+        pytest.warns(FutureWarning, match=msg_re("connectivities")),
+        pytest.warns(ad.OldFormatWarning),
+    ):
         ad.read_h5ad(pth)
 
-    assert any(issubclass(w.category, ad.OldFormatWarning) for w in record), [
-        w.message for w in record if not issubclass(w.category, FutureWarning)
-    ]
-    assert any(
-        issubclass(w.category, FutureWarning)
-        and re.match(
-            r"Moving element from \.uns\['neighbors']\['distances'] to \.obsp\['distances']\.",
-            str(w.message),
-        )
-        for w in record
-    ), [w.message for w in record if not issubclass(w.category, ad.OldFormatWarning)]
 
-
-def test_old_format_warning_not_thrown(tmp_path):
+def test_old_format_warning_not_thrown(tmp_path: Path) -> None:
     pth = tmp_path / "current.h5ad"
-    adata = gen_adata((20, 10))
+    adata = gen_adata((20, 10), **GEN_ADATA_NO_XARRAY_ARGS)
     adata.write_h5ad(pth)
 
     with warnings.catch_warnings(record=True) as record:
