@@ -3,6 +3,7 @@ from __future__ import annotations
 from contextlib import ExitStack
 from copy import deepcopy
 from operator import mul
+from typing import TYPE_CHECKING
 
 import joblib
 import numpy as np
@@ -21,7 +22,7 @@ from anndata._core.views import (
     SparseCSRArrayView,
     SparseCSRMatrixView,
 )
-from anndata.compat import CAN_USE_SPARSE_ARRAY, CupyCSCMatrix, DaskArray
+from anndata.compat import CupyCSCMatrix, DaskArray
 from anndata.tests.helpers import (
     BASE_MATRIX_PARAMS,
     CUPY_MATRIX_PARAMS,
@@ -34,6 +35,9 @@ from anndata.tests.helpers import (
     subset_func,
 )
 from anndata.utils import asarray
+
+if TYPE_CHECKING:
+    from types import EllipsisType
 
 IGNORE_SPARSE_EFFICIENCY_WARNING = pytest.mark.filterwarnings(
     "ignore:Changing the sparsity structure:scipy.sparse.SparseEfficiencyWarning"
@@ -148,7 +152,7 @@ def test_view_subset_shapes():
 
     view = adata[:, ::2]
     assert view.var.shape == (5, 8)
-    assert {k: v.shape[0] for k, v in view.varm.items()} == {k: 5 for k in view.varm}
+    assert {k: v.shape[0] for k, v in view.varm.items()} == dict.fromkeys(view.varm, 5)
 
 
 def test_modify_view_component(matrix_type, mapping_name, request):
@@ -175,7 +179,7 @@ def test_modify_view_component(matrix_type, mapping_name, request):
 
     assert init_hash == hash_func(adata)
 
-    if "sparse_array_dask_array" in request.node.callspec.id and CAN_USE_SPARSE_ARRAY:
+    if "sparse_array_dask_array" in request.node.callspec.id:
         msg = "sparse arrays in dask are generally expected to fail but in this case they do not"
         pytest.fail(msg)
 
@@ -326,7 +330,7 @@ def test_not_set_subset_X(matrix_type_base, subset_func):
     with pytest.warns(ad.ImplicitModificationWarning, match=r".*X.*"):
         subset.X[:, internal_idx] = 1
     assert not subset.is_view
-    assert not np.any(asarray(adata.X != orig_X_val))
+    assert not np.any(asarray(orig_X_val != adata.X))
 
     assert init_hash == joblib.hash(adata)
     assert isinstance(subset.X, type(adata.X))
@@ -354,7 +358,7 @@ def test_not_set_subset_X_dask(matrix_type_no_gpu, subset_func):
     with pytest.warns(ad.ImplicitModificationWarning, match=r".*X.*"):
         subset.X[:, internal_idx] = 1
     assert not subset.is_view
-    assert not np.any(asarray(adata.X != orig_X_val))
+    assert not np.any(asarray(orig_X_val != adata.X))
 
     assert init_hash == tokenize(adata)
     assert isinstance(subset.X, type(adata.X))
@@ -530,7 +534,7 @@ def test_view_of_view(matrix_type, subset_func, subset_func2):
         pytest.xfail("Other subset generating functions have trouble with this")
     var_s1 = subset_func(adata.var_names, min_size=4)
     var_view1 = adata[:, var_s1]
-    adata[:, var_s1].X
+    adata[:, var_s1].X  # noqa: B018
     var_s2 = subset_func2(var_view1.var_names)
     var_view2 = var_view1[:, var_s2]
     assert var_view2._adata_ref is adata
@@ -627,7 +631,7 @@ def test_invalid_scalar_index(adata, index):
 
 @pytest.mark.parametrize("obs", [False, True])
 @pytest.mark.parametrize("index", [-100, -50, -1])
-def test_negative_scalar_index(adata, index: int, obs: bool):
+def test_negative_scalar_index(*, adata, index: int, obs: bool):
     pos_index = index + (adata.n_obs if obs else adata.n_vars)
 
     if obs:
@@ -669,9 +673,7 @@ def test_viewness_propagation_allclose(adata):
     assert np.allclose(a.varm["o"], b.varm["o"].copy(), equal_nan=True)
 
 
-spmat = [sparse.csr_matrix, sparse.csc_matrix]
-if CAN_USE_SPARSE_ARRAY:
-    spmat += [sparse.csr_array, sparse.csc_array]
+spmat = [sparse.csr_matrix, sparse.csc_matrix, sparse.csr_array, sparse.csc_array]
 
 
 @pytest.mark.parametrize("spmat", spmat)
@@ -689,10 +691,7 @@ def test_deepcopy_subset(adata, spmat: type):
     view_type = (
         SparseCSRMatrixView if spmat is sparse.csr_matrix else SparseCSCMatrixView
     )
-    if CAN_USE_SPARSE_ARRAY:
-        view_type = (
-            SparseCSRArrayView if spmat is sparse.csr_array else SparseCSCArrayView
-        )
+    view_type = SparseCSRArrayView if spmat is sparse.csr_array else SparseCSCArrayView
     assert not isinstance(
         adata.obsp["spmat"],
         view_type,
@@ -700,9 +699,13 @@ def test_deepcopy_subset(adata, spmat: type):
     np.testing.assert_array_equal(adata.obsp["spmat"].shape, (10, 10))
 
 
-array_type = [asarray, sparse.csr_matrix, sparse.csc_matrix]
-if CAN_USE_SPARSE_ARRAY:
-    array_type += [sparse.csr_array, sparse.csc_array]
+array_type = [
+    asarray,
+    sparse.csr_matrix,
+    sparse.csc_matrix,
+    sparse.csr_array,
+    sparse.csc_array,
+]
 
 
 # https://github.com/scverse/anndata/issues/680
@@ -722,12 +725,7 @@ def test_view_mixin_copies_data(adata, array_type: type, attr):
         getattr(adata, attr)["arr"] = X
 
     view = adata[:50]
-
-    if attr == "X":
-        arr_view = view.X
-    else:
-        arr_view = getattr(view, attr)["arr"]
-
+    arr_view = view.X if attr == "X" else getattr(view, attr)["arr"]
     arr_view_copy = arr_view.copy()
 
     if sparse.issparse(X):
@@ -784,6 +782,46 @@ def test_dataframe_view_index_setting():
     assert isinstance(a2.obs, pd.DataFrame)
     assert a1.obs.index.values.tolist() == ["aa", "bb"]
     assert a2.obs.index.values.tolist() == ["a", "b"]
+
+
+def test_ellipsis_index(
+    ellipsis_index: tuple[EllipsisType | slice, ...] | EllipsisType,
+    equivalent_ellipsis_index: tuple[slice, slice],
+    matrix_type,
+):
+    adata = gen_adata((10, 10), X_type=matrix_type, **GEN_ADATA_DASK_ARGS)
+    subset_ellipsis = adata[ellipsis_index]
+    subset = adata[equivalent_ellipsis_index]
+    assert_equal(subset_ellipsis, subset)
+
+
+@pytest.mark.parametrize(
+    ("index", "expected_error"),
+    [
+        ((..., 0, ...), r"only have a single ellipsis"),
+        ((0, 0, 0), r"Received a length 3 index"),
+    ],
+    ids=["ellipsis-int-ellipsis", "int-int-int"],
+)
+def test_index_3d_errors(index: tuple[int | EllipsisType, ...], expected_error: str):
+    with pytest.raises(IndexError, match=expected_error):
+        gen_adata((10, 10))[index]
+
+
+@pytest.mark.parametrize(
+    "index",
+    [
+        pytest.param(sparse.csr_matrix(np.random.random((1, 10))), id="sparse"),
+        pytest.param([1.2, 3.4], id="list"),
+        *(
+            pytest.param(np.array([1.2, 2.3], dtype=dtype), id=f"ndarray-{dtype}")
+            for dtype in [np.float32, np.float64]
+        ),
+    ],
+)
+def test_index_float_sequence_raises_error(index):
+    with pytest.raises(IndexError, match=r"has floating point values"):
+        gen_adata((10, 10))[index]
 
 
 # @pytest.mark.parametrize("dim", ["obs", "var"])
